Stock prices for Facebook from 2013-2016

A dataset containing the date, open, high, low, close, volume, and adjusted stock prices for Facebook from 2013-2016.

FB

Format

A tibble with 1,008 rows and 8 variables:

symbol

stock ticker symbol

date

trade date

open

stock price at the open of trading, in USD

high

stock price at the highest point during trading, in USD

low

stock price at the lowest point during trading, in USD

close

stock price at the close of trading, in USD

volume

number of shares traded

adjusted

stock price at the close of trading adjusted for stock splits, in USD

Source

https://www.investopedia.com/terms/f/fang-stocks-fb-amzn.asp