Change tbl_time periodicity

Convert a tbl_time object from daily to monthly, from minute data to hourly, and more. This allows the user to easily aggregate data to a less granular level by taking the value from either the beginning or end of the period.

as_period(.tbl_time, period = "yearly", start_date = NULL,
  side = "start", include_endpoints = FALSE, ...)

Arguments

.tbl_time
A tbl_time object.

period
A character specification used for time-based grouping. The general format to use is "frequency period" where frequency is a number like 1 or 2, and period is an interval like weekly or yearly. There must be a space between the two.

Note that you can pass the specification in a flexible way:

1 Year: '1 year' / '1 Y' / '1 yearly' / 'yearly'

This shorthand is available for year, quarter, month, day, hour, minute, second, millisecond and microsecond periodicities.

Additionally, you have the option of passing in a vector of dates to use as custom and more flexible boundaries.

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

side
Whether to return the date at the beginning or the end of the new period. By default, the "start" of the period. Use "end" to change to the end of the period.

include_endpoints
Whether to include the first or last data point in addition to the transformed data.

...
Not currently used.

Details

This function respects dplyr::group_by() groups.

The side argument is useful when you want to return data at, say, the end of a quarter, or the end of a month.

include_endpoints can be useful when calculating a change over time. In addition to changing to monthly dates, you often need the first data point as a baseline for the first calculation.

Examples

# Basic usage ---------------------------------------------------------------

# FB stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Aggregate FB to yearly data
as_period(FB, "yearly")

#> # A time tibble: 4 x 8
#> [90m# Index: date[39m
#>   symbol date        open  high   low close   volume adjusted
#>   [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m2[39m FB     2014-01-02  54.8  55.2  54.2  54.7 43[4m1[24m[4m9[24m[4m5[24m500     54.7
#> [90m3[39m FB     2015-01-02  78.6  78.9  77.7  78.4 18[4m1[24m[4m7[24m[4m7[24m500     78.4
#> [90m4[39m FB     2016-01-04 102.  102.   99.8 102.  37[4m9[24m[4m1[24m[4m2[24m400    102. 

# Aggregate FB to every 2 years
as_period(FB, "2 years")

#> # A time tibble: 3 x 8
#> [90m# Index: date[39m
#>   symbol date        open  high   low close   volume adjusted
#>   [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m2[39m FB     2014-01-02  54.8  55.2  54.2  54.7 43[4m1[24m[4m9[24m[4m5[24m500     54.7
#> [90m3[39m FB     2016-01-04 102.  102.   99.8 102.  37[4m9[24m[4m1[24m[4m2[24m400    102. 

# Aggregate FB to yearly data, but use the last data point available
# in that period
as_period(FB, "yearly", side = "end")

#> # A time tibble: 4 x 8
#> [90m# Index: date[39m
#>   symbol date        open  high   low close   volume adjusted
#>   [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m FB     2013-12-31  54.1  54.9  53.9  54.7 43[4m0[24m[4m7[24m[4m6[24m200     54.7
#> [90m2[39m FB     2014-12-31  79.5  79.8  77.9  78.0 19[4m9[24m[4m3[24m[4m5[24m400     78.0
#> [90m3[39m FB     2015-12-31 106   106.  105.  105.  18[4m2[24m[4m9[24m[4m8[24m700    105. 
#> [90m4[39m FB     2016-12-30 117.  117.  115.  115.  18[4m6[24m[4m0[24m[4m0[24m100    115. 

# Aggregate FB to yearly data, end of period, and include the first
# endpoint
as_period(FB, "yearly", side = "end", include_endpoints = TRUE)

#> # A time tibble: 5 x 8
#> [90m# Index: date[39m
#>   symbol date        open  high   low close   volume adjusted
#>   [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m2[39m FB     2013-12-31  54.1  54.9  53.9  54.7 43[4m0[24m[4m7[24m[4m6[24m200     54.7
#> [90m3[39m FB     2014-12-31  79.5  79.8  77.9  78.0 19[4m9[24m[4m3[24m[4m5[24m400     78.0
#> [90m4[39m FB     2015-12-31 106   106.  105.  105.  18[4m2[24m[4m9[24m[4m8[24m700    105. 
#> [90m5[39m FB     2016-12-30 117.  117.  115.  115.  18[4m6[24m[4m0[24m[4m0[24m100    115. 

# Aggregate to weekly. Notice that it only uses the earliest day available
# in the data set at that periodicity. It will not set the date of the first
# row to 2013-01-01 because that date did not exist in the original data set.
as_period(FB, "weekly")

#> # A time tibble: 209 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close   volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-07  28.7  29.8  28.6  29.4 83[4m7[24m[4m8[24m[4m1[24m800     29.4
#> [90m 3[39m FB     2013-01-14  32.1  32.2  30.6  31.0 98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m 4[39m FB     2013-01-22  29.8  30.9  29.7  30.7 55[4m2[24m[4m4[24m[4m3[24m300     30.7
#> [90m 5[39m FB     2013-01-28  31.9  32.5  31.8  32.5 59[4m6[24m[4m8[24m[4m2[24m500     32.5
#> [90m 6[39m FB     2013-02-04  29.1  29.2  28.0  28.1 92[4m3[24m[4m6[24m[4m2[24m200     28.1
#> [90m 7[39m FB     2013-02-11  28.6  28.7  28.0  28.3 37[4m3[24m[4m6[24m[4m1[24m800     28.3
#> [90m 8[39m FB     2013-02-19  28.2  29.1  28.1  28.9 49[4m3[24m[4m9[24m[4m6[24m400     28.9
#> [90m 9[39m FB     2013-02-25  27.2  27.6  27.2  27.3 34[4m6[24m[4m5[24m[4m2[24m000     27.3
#> [90m10[39m FB     2013-03-04  27.8  28.1  27.4  27.7 32[4m4[24m[4m0[24m[4m0[24m700     27.7
#> [90m# … with 199 more rows[39m

# Aggregate to every other week
as_period(FB, "2 weeks")

#> Warning: Multi-unit not supported for weeks. Ignoring.

#> Warning: Multi-unit not supported for weeks. Ignoring.

#> # A time tibble: 105 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close   volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-14  32.1  32.2  30.6  31.0 98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m 3[39m FB     2013-01-28  31.9  32.5  31.8  32.5 59[4m6[24m[4m8[24m[4m2[24m500     32.5
#> [90m 4[39m FB     2013-02-11  28.6  28.7  28.0  28.3 37[4m3[24m[4m6[24m[4m1[24m800     28.3
#> [90m 5[39m FB     2013-02-25  27.2  27.6  27.2  27.3 34[4m6[24m[4m5[24m[4m2[24m000     27.3
#> [90m 6[39m FB     2013-03-11  28.0  28.6  27.8  28.1 35[4m6[24m[4m4[24m[4m2[24m100     28.1
#> [90m 7[39m FB     2013-03-25  25.8  25.8  25.1  25.1 39[4m1[24m[4m9[24m[4m9[24m000     25.1
#> [90m 8[39m FB     2013-04-08  27.2  27.2  26.6  26.8 27[4m2[24m[4m5[24m[4m6[24m000     26.8
#> [90m 9[39m FB     2013-04-22  25.8  26.4  25.7  26.0 25[4m6[24m[4m8[24m[4m7[24m600     26.0
#> [90m10[39m FB     2013-05-06  28.3  28.5  27.5  27.6 43[4m9[24m[4m3[24m[4m9[24m400     27.6
#> [90m# … with 95 more rows[39m

# FB is daily data, aggregate to minute?
# Not allowed for Date class indices, an error is thrown
# as_period(FB, "minute")

# Grouped usage -------------------------------------------------------------

# FANG contains Facebook, Amazon, Netflix and Google stock prices
data(FANG)
FANG <- as_tbl_time(FANG, date)

FANG <- dplyr::group_by(FANG, symbol)

# Respects groups
as_period(FANG, "yearly")

#> # A time tibble: 16 x 8
#> [90m# Index:  date[39m
#> [90m# Groups: symbol [4][39m
#>    symbol date         open   high    low  close   volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m      [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02   27.4   28.2   27.4   28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2014-01-02   54.8   55.2   54.2   54.7 43[4m1[24m[4m9[24m[4m5[24m500     54.7
#> [90m 3[39m FB     2015-01-02   78.6   78.9   77.7   78.4 18[4m1[24m[4m7[24m[4m7[24m500     78.4
#> [90m 4[39m FB     2016-01-04  102.   102.    99.8  102.  37[4m9[24m[4m1[24m[4m2[24m400    102. 
#> [90m 5[39m AMZN   2013-01-02  256.   258.   253.   257.   3[4m2[24m[4m7[24m[4m1[24m000    257. 
#> [90m 6[39m AMZN   2014-01-02  399.   399.   394.   398.   2[4m1[24m[4m3[24m[4m7[24m800    398. 
#> [90m 7[39m AMZN   2015-01-02  313.   315.   307.   309.   2[4m7[24m[4m8[24m[4m3[24m200    309. 
#> [90m 8[39m AMZN   2016-01-04  656.   658.   628.   637.   9[4m3[24m[4m1[24m[4m4[24m500    637. 
#> [90m 9[39m NFLX   2013-01-02   95.2   95.8   90.7   92.0 19[4m4[24m[4m3[24m[4m1[24m300     13.1
#> [90m10[39m NFLX   2014-01-02  367.   368.   361.   363.  12[4m3[24m[4m2[24m[4m5[24m600     51.8
#> [90m11[39m NFLX   2015-01-02  344.   352.   341.   349.  13[4m4[24m[4m7[24m[4m5[24m000     49.8
#> [90m12[39m NFLX   2016-01-04  109    110    105.   110.  20[4m7[24m[4m9[24m[4m4[24m800    110. 
#> [90m13[39m GOOG   2013-01-02  719.   727.   717.   723.   5[4m1[24m[4m0[24m[4m1[24m500    361. 
#> [90m14[39m GOOG   2014-01-02 [4m1[24m115.  [4m1[24m118.  [4m1[24m108.  [4m1[24m113.   3[4m6[24m[4m5[24m[4m6[24m400    556. 
#> [90m15[39m GOOG   2015-01-02  529.   531.   524.   525.   1[4m4[24m[4m4[24m[4m7[24m500    525. 
#> [90m16[39m GOOG   2016-01-04  743    744.   731.   742.   3[4m2[24m[4m7[24m[4m2[24m800    742. 

# Every 6 months, respecting groups
as_period(FANG, "6 months")

#> # A time tibble: 32 x 8
#> [90m# Index:  date[39m
#> [90m# Groups: symbol [4][39m
#>    symbol date        open  high   low close   volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-07-01  25.0  25.1  24.6  24.8 20[4m5[24m[4m8[24m[4m2[24m200     24.8
#> [90m 3[39m FB     2014-01-02  54.8  55.2  54.2  54.7 43[4m1[24m[4m9[24m[4m5[24m500     54.7
#> [90m 4[39m FB     2014-07-01  67.6  68.4  67.4  68.1 33[4m2[24m[4m4[24m[4m3[24m000     68.1
#> [90m 5[39m FB     2015-01-02  78.6  78.9  77.7  78.4 18[4m1[24m[4m7[24m[4m7[24m500     78.4
#> [90m 6[39m FB     2015-07-01  86.8  87.9  86.5  86.9 25[4m2[24m[4m6[24m[4m0[24m000     86.9
#> [90m 7[39m FB     2016-01-04 102.  102.   99.8 102.  37[4m9[24m[4m1[24m[4m2[24m400    102. 
#> [90m 8[39m FB     2016-07-01 114.  115.  114.  114.  14[4m9[24m[4m8[24m[4m0[24m000    114. 
#> [90m 9[39m AMZN   2013-01-02 256.  258.  253.  257.   3[4m2[24m[4m7[24m[4m1[24m000    257. 
#> [90m10[39m AMZN   2013-07-01 279   283.  277.  282.   2[4m8[24m[4m8[24m[4m8[24m200    282. 
#> [90m# … with 22 more rows[39m

# Using start_date ----------------------------------------------------------

#### One method using start_date

# FB stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# The Facebook series starts at 2013-01-02 so the 'every 2 day' counter
# starts at that date as well. Groups become (2013-01-02, 2013-01-03),
# (2013-01-04, 2013-01-05) and so on.
as_period(FB, "2 day")

#> # A time tibble: 619 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close    volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8
#> [90m 3[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4
#> [90m 4[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6
#> [90m 5[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7
#> [90m 6[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m 7[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1
#> [90m 8[39m FB     2013-01-17  30.1  30.4  30.0  30.1  40[4m2[24m[4m5[24m[4m6[24m700     30.1
#> [90m 9[39m FB     2013-01-22  29.8  30.9  29.7  30.7  55[4m2[24m[4m4[24m[4m3[24m300     30.7
#> [90m10[39m FB     2013-01-23  31.1  31.5  30.8  30.8  48[4m8[24m[4m9[24m[4m9[24m800     30.8
#> [90m# … with 609 more rows[39m

# Specifying the `start_date = "2013-01-01"` might be preferable.
# Groups become (2013-01-01, 2013-01-02), (2013-01-03, 2013-01-04) and so on.
as_period(FB, "2 day", start_date = "2013-01-01")

#> # A time tibble: 619 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close    volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8
#> [90m 3[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4
#> [90m 4[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6
#> [90m 5[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7
#> [90m 6[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m 7[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1
#> [90m 8[39m FB     2013-01-17  30.1  30.4  30.0  30.1  40[4m2[24m[4m5[24m[4m6[24m700     30.1
#> [90m 9[39m FB     2013-01-22  29.8  30.9  29.7  30.7  55[4m2[24m[4m4[24m[4m3[24m300     30.7
#> [90m10[39m FB     2013-01-23  31.1  31.5  30.8  30.8  48[4m8[24m[4m9[24m[4m9[24m800     30.8
#> [90m# … with 609 more rows[39m

#### Equivalent method using an index vector

# FB stock prices
data(FB)
FB <- as_tbl_time(FB, date)

custom_period <- create_series(
  time_formula = dplyr::first(FB$date) - 1 ~ dplyr::last(FB$date),
  period       = "2 day",
  class        = "Date",
  as_vector    = TRUE)

FB %>%
  as_tbl_time(date) %>%
  as_period(period = custom_period)

#> # A time tibble: 619 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close    volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8
#> [90m 3[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4
#> [90m 4[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6
#> [90m 5[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7
#> [90m 6[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m 7[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1
#> [90m 8[39m FB     2013-01-17  30.1  30.4  30.0  30.1  40[4m2[24m[4m5[24m[4m6[24m700     30.1
#> [90m 9[39m FB     2013-01-22  29.8  30.9  29.7  30.7  55[4m2[24m[4m4[24m[4m3[24m300     30.7
#> [90m10[39m FB     2013-01-23  31.1  31.5  30.8  30.8  48[4m8[24m[4m9[24m[4m9[24m800     30.8
#> [90m# … with 609 more rows[39m

# Manually calculating returns at different periods -------------------------

data(FB)

# Annual Returns
# Convert to end of year periodicity, but include the endpoints to use as
# a reference for the first return calculation. Then calculate returns.
FB %>%
  as_tbl_time(date) %>%
  as_period("1 y", side = "end", include_endpoints = TRUE) %>%
  dplyr::mutate(yearly_return = adjusted / dplyr::lag(adjusted) - 1)

#> # A time tibble: 5 x 9
#> [90m# Index: date[39m
#>   symbol date        open  high   low close   volume adjusted yearly_return
#>   [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m         [3m[90m<dbl>[39m[23m
#> [90m1[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28         [31mNA[39m     
#> [90m2[39m FB     2013-12-31  54.1  54.9  53.9  54.7 43[4m0[24m[4m7[24m[4m6[24m200     54.7        0.952 
#> [90m3[39m FB     2014-12-31  79.5  79.8  77.9  78.0 19[4m9[24m[4m3[24m[4m5[24m400     78.0        0.428 
#> [90m4[39m FB     2015-12-31 106   106.  105.  105.  18[4m2[24m[4m9[24m[4m8[24m700    105.         0.341 
#> [90m5[39m FB     2016-12-30 117.  117.  115.  115.  18[4m6[24m[4m0[24m[4m0[24m100    115.         0.099[4m3[24m