Create a tbl_time object with a sequence of regularly spaced dates

create_series() allows the user to quickly create a tbl_time object with a date column populated with a sequence of dates.

create_series(time_formula, period = "daily", include_end = FALSE,
  tz = NULL, force_class = NULL, as_date_vector = FALSE)

Arguments

time_formula
A period to create the series over. This is specified as a formula. See the Details section of time_filter() for more information.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

include_end
Whether to always include the RHS of the time_formula even if it does not match the regularly spaced index.

tz
Time zone of the new series.

force_class
Either "Date" or "POSIXct". The default is to infer the simplest class required from the period specified, but this will override that.

as_date_vector
Should the series be returned as a vector instead of a tibble?

Examples

# Every day in 2013
create_series(~2013, "daily")

#> # A time tibble: 365 x 1
#> # Index: date
#>          date
#>        <date>
#>  1 2013-01-01
#>  2 2013-01-02
#>  3 2013-01-03
#>  4 2013-01-04
#>  5 2013-01-05
#>  6 2013-01-06
#>  7 2013-01-07
#>  8 2013-01-08
#>  9 2013-01-09
#> 10 2013-01-10
#> # ... with 355 more rows

# Every other day in 2013
create_series(~2013, 2~d)

#> # A time tibble: 183 x 1
#> # Index: date
#>          date
#>        <date>
#>  1 2013-01-01
#>  2 2013-01-03
#>  3 2013-01-05
#>  4 2013-01-07
#>  5 2013-01-09
#>  6 2013-01-11
#>  7 2013-01-13
#>  8 2013-01-15
#>  9 2013-01-17
#> 10 2013-01-19
#> # ... with 173 more rows

# Every quarter in 2013
create_series(~2013, 1~q)

#> # A time tibble: 4 x 1
#> # Index: date
#>         date
#>       <date>
#> 1 2013-01-01
#> 2 2013-04-01
#> 3 2013-07-01
#> 4 2013-10-01

# Daily series for 2 years
create_series(2013~2015, 1~d)

#> # A time tibble: 1,095 x 1
#> # Index: date
#>          date
#>        <date>
#>  1 2013-01-01
#>  2 2013-01-02
#>  3 2013-01-03
#>  4 2013-01-04
#>  5 2013-01-05
#>  6 2013-01-06
#>  7 2013-01-07
#>  8 2013-01-08
#>  9 2013-01-09
#> 10 2013-01-10
#> # ... with 1,085 more rows

# Minute series for 2 months
create_series(2012-01~2012-02, 1~M)

#> # A time tibble: 86,400 x 1
#> # Index: date
#>                   date
#>                 <dttm>
#>  1 2012-01-01 00:00:00
#>  2 2012-01-01 00:01:00
#>  3 2012-01-01 00:02:00
#>  4 2012-01-01 00:03:00
#>  5 2012-01-01 00:04:00
#>  6 2012-01-01 00:05:00
#>  7 2012-01-01 00:06:00
#>  8 2012-01-01 00:07:00
#>  9 2012-01-01 00:08:00
#> 10 2012-01-01 00:09:00
#> # ... with 86,390 more rows

# Second series for 2 minutes
create_series(2011-01-01 + 12:10:00 ~ 2011-01-01 + 12:12:00, 1~s)

#> # A time tibble: 121 x 1
#> # Index: date
#>                   date
#>                 <dttm>
#>  1 2011-01-01 12:10:00
#>  2 2011-01-01 12:10:01
#>  3 2011-01-01 12:10:02
#>  4 2011-01-01 12:10:03
#>  5 2011-01-01 12:10:04
#>  6 2011-01-01 12:10:05
#>  7 2011-01-01 12:10:06
#>  8 2011-01-01 12:10:07
#>  9 2011-01-01 12:10:08
#> 10 2011-01-01 12:10:09
#> # ... with 111 more rows

# Force POSIXct class
create_series(~2013, 1~d, force_class = "POSIXct")

#> # A time tibble: 365 x 1
#> # Index: date
#>          date
#>        <dttm>
#>  1 2013-01-01
#>  2 2013-01-02
#>  3 2013-01-03
#>  4 2013-01-04
#>  5 2013-01-05
#>  6 2013-01-06
#>  7 2013-01-07
#>  8 2013-01-08
#>  9 2013-01-09
#> 10 2013-01-10
#> # ... with 355 more rows