Parse a character period specification

The period is parsed into frequency and period and returned as a named list.

parse_period(period)

Arguments

period
A character specification used for time-based grouping. The general format to use is "frequency period" where frequency is a number like 1 or 2, and period is an interval like weekly or yearly. There must be a space between the two.

Note that you can pass the specification in a flexible way:

1 Year: '1 year' / '1 Y' / '1 yearly' / 'yearly'

This shorthand is available for year, quarter, month, day, hour, minute, second, millisecond and microsecond periodicities.

Additionally, you have the option of passing in a vector of dates to use as custom and more flexible boundaries.

Examples

parse_period('2 day')

#> $freq
#> [1] 2
#> 
#> $period
#> [1] "day"
#>