Create a rolling version of any function

rollify returns a rolling version of the input function, with a rolling window specified by the user.

rollify(.f, window = 1, unlist = TRUE, na_value = NULL)

Arguments

.f
A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

window
The window size to roll over

unlist
If the function returns a single value each time it is called, use unlist = TRUE. If the function returns more than one value, or a more complicated object (like a linear model), use unlist = FALSE to create a list-column of the rolling results.

na_value
A default value for the NA values at the beginning of the roll.

Details

The intended use of rollify is to turn a function into a rolling version of itself for use inside of a call to dplyr::mutate(), however it works equally as well when called from purrr::map().

Because of it's intended use with dplyr::mutate(), rollify creates a function that always returns output with the same length of the input, aligned right, and filled with NA unless otherwise specified by na_value.

The form of the .f argument is the same as the form that can be passed to purrr::map(). Use .x or . to refer to the first object to roll over, and .y to refer to the second object if required. The examples explain this further.

If optional arguments to the function are required, specify them in the call to rollify, and not in the call to the rolling version of the function. See the examples for more details.

See also

Examples

# Rolling mean --------------------------------------------------------------

data(FB)

# Turn the normal mean function into a rolling mean with a 5 row window
mean_roll_5 <- rollify(mean, window = 5)

dplyr::mutate(FB,
       normal_mean  = mean(adjusted),
       rolling_mean = mean_roll_5(adjusted))

#> [90m# A tibble: 1,008 x 10[39m
#>    symbol date        open  high   low close volume adjusted normal_mean
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m       [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28   6.98[90me[39m7     28          77.5
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8 6.31[90me[39m7     27.8        77.5
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8 7.27[90me[39m7     28.8        77.5
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4 8.38[90me[39m7     29.4        77.5
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1 4.59[90me[39m7     29.1        77.5
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 1.05[90me[39m8     30.6        77.5
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3 9.53[90me[39m7     31.3        77.5
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7 8.96[90me[39m7     31.7        77.5
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0 9.89[90me[39m7     31.0        77.5
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 1.73[90me[39m8     30.1        77.5
#> [90m# … with 998 more rows, and 1 more variable: rolling_mean [3m[90m<dbl>[90m[23m[39m

# There's nothing stopping you from combining multiple rolling functions with
# different window sizes in the same mutate call
mean_roll_10 <- rollify(mean, window = 10)

dplyr::mutate(FB,
       rolling_mean_5  = mean_roll_5(adjusted),
       rolling_mean_10 = mean_roll_10(adjusted))

#> [90m# A tibble: 1,008 x 10[39m
#>    symbol date        open  high   low close volume adjusted rolling_mean_5
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m          [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28   6.98[90me[39m7     28             [31mNA[39m  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8 6.31[90me[39m7     27.8           [31mNA[39m  
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8 7.27[90me[39m7     28.8           [31mNA[39m  
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4 8.38[90me[39m7     29.4           [31mNA[39m  
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1 4.59[90me[39m7     29.1           28.6
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 1.05[90me[39m8     30.6           29.1
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3 9.53[90me[39m7     31.3           29.8
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7 8.96[90me[39m7     31.7           30.4
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0 9.89[90me[39m7     31.0           30.7
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 1.73[90me[39m8     30.1           30.9
#> [90m# … with 998 more rows, and 1 more variable: rolling_mean_10 [3m[90m<dbl>[90m[23m[39m

# Functions with multiple args and optional args ----------------------------

# With 2 args, use the purrr syntax of ~ and .x, .y
# Rolling correlation example
cor_roll <- rollify(~cor(.x, .y), window = 5)

dplyr::mutate(FB, running_cor = cor_roll(adjusted, open))

#> [90m# A tibble: 1,008 x 9[39m
#>    symbol date        open  high   low close    volume adjusted running_cor
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m       [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28        [31mNA[39m    
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8      [31mNA[39m    
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8  72[4m7[24m[4m1[24m[4m5[24m400     28.8      [31mNA[39m    
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4      [31mNA[39m    
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1  45[4m8[24m[4m7[24m[4m1[24m300     29.1       0.749
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6       0.805
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3  95[4m3[24m[4m1[24m[4m6[24m400     31.3       0.859
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7       0.884
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0       0.667
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1       0.379
#> [90m# … with 998 more rows[39m

# With >2 args, create an anonymous function with >2 args or use
# the purrr convention of ..1, ..2, ..3 to refer to the arguments
avg_of_avgs <- rollify(function(x, y, z) {
                         (mean(x) + mean(y) + mean(z)) / 3
                       },
                       window = 10)

# Or
avg_of_avgs <- rollify(~(mean(..1) + mean(..2) + mean(..3)) / 3,
                       window = 10)

dplyr::mutate(FB, avg_of_avgs = avg_of_avgs(open, high, low))

#> [90m# A tibble: 1,008 x 9[39m
#>    symbol date        open  high   low close    volume adjusted avg_of_avgs
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m       [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28          [31mNA[39m  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8        [31mNA[39m  
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8  72[4m7[24m[4m1[24m[4m5[24m400     28.8        [31mNA[39m  
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4        [31mNA[39m  
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1  45[4m8[24m[4m7[24m[4m1[24m300     29.1        [31mNA[39m  
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6        [31mNA[39m  
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3  95[4m3[24m[4m1[24m[4m6[24m400     31.3        [31mNA[39m  
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7        [31mNA[39m  
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0        [31mNA[39m  
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1        29.7
#> [90m# … with 998 more rows[39m

# Optional arguments MUST be passed at the creation of the rolling function
# Only data arguments that are "rolled over" are allowed when calling the
# rolling version of the function
FB$adjusted[1] <- NA

roll_mean_na_rm <- rollify(~mean(.x, na.rm = TRUE), window = 5)

dplyr::mutate(FB, roll_mean = roll_mean_na_rm(adjusted))

#> [90m# A tibble: 1,008 x 9[39m
#>    symbol date        open  high   low close    volume adjusted roll_mean
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     [31mNA[39m        [31mNA[39m  
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8      [31mNA[39m  
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8  72[4m7[24m[4m1[24m[4m5[24m400     28.8      [31mNA[39m  
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4      [31mNA[39m  
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1  45[4m8[24m[4m7[24m[4m1[24m300     29.1      28.8
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6      29.1
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3  95[4m3[24m[4m1[24m[4m6[24m400     31.3      29.8
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7      30.4
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0      30.7
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1      30.9
#> [90m# … with 998 more rows[39m

# Returning multiple values -------------------------------------------------

data(FB)

# If the function returns >1 value, set the `unlist = FALSE` argument
# Running 5 number summary
summary_roll <- rollify(summary, window = 5, unlist = FALSE)

FB_summarised <- dplyr::mutate(FB, summary_roll = summary_roll(adjusted))
FB_summarised$summary_roll[[5]]

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>   27.77   28.51   28.91   28.75   29.15   29.42       1 

# dplyr::bind_rows() is often helpful in these cases to get
# meaningful output

summary_roll <- rollify(~dplyr::bind_rows(summary(.)), window = 5, unlist = FALSE)
FB_summarised <- dplyr::mutate(FB, summary_roll = summary_roll(adjusted))
FB_summarised %>%
  dplyr::filter(!is.na(summary_roll)) %>%
  tidyr::unnest(summary_roll)

#> [90m# A tibble: 1,004 x 15[39m
#>    symbol date        open  high   low close volume adjusted  Min. `1st Qu.`
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-08  29.5  29.6  28.9  29.1 4.59[90me[39m7     29.1  27.8      28.5
#> [90m 2[39m FB     2013-01-09  29.7  30.6  29.5  30.6 1.05[90me[39m8     30.6  27.8      28.8
#> [90m 3[39m FB     2013-01-10  30.6  31.5  30.3  31.3 9.53[90me[39m7     31.3  28.8      29.1
#> [90m 4[39m FB     2013-01-11  31.3  32.0  31.1  31.7 8.96[90me[39m7     31.7  29.1      29.4
#> [90m 5[39m FB     2013-01-14  32.1  32.2  30.6  31.0 9.89[90me[39m7     31.0  29.1      30.6
#> [90m 6[39m FB     2013-01-15  30.6  31.7  29.9  30.1 1.73[90me[39m8     30.1  30.1      30.6
#> [90m 7[39m FB     2013-01-16  30.2  30.4  29.5  29.8 7.53[90me[39m7     29.8  29.8      30.1
#> [90m 8[39m FB     2013-01-17  30.1  30.4  30.0  30.1 4.03[90me[39m7     30.1  29.8      30.1
#> [90m 9[39m FB     2013-01-18  30.3  30.4  29.3  29.7 4.96[90me[39m7     29.7  29.7      29.8
#> [90m10[39m FB     2013-01-22  29.8  30.9  29.7  30.7 5.52[90me[39m7     30.7  29.7      29.8
#> [90m# … with 994 more rows, and 5 more variables: Median [3m[90m<dbl>[90m[23m, Mean [3m[90m<dbl>[90m[23m, `3rd
#> #   Qu.` [3m[90m<dbl>[90m[23m, Max. [3m[90m<dbl>[90m[23m, `NA's` [3m[90m<dbl>[90m[23m[39m

# Rolling regressions -------------------------------------------------------

# Extending an example from R 4 Data Science on "Many Models".
# For each country in the gapminder data, calculate a linear regression
# every 5 periods of lifeExp ~ year
library(gapminder)

# Rolling regressions are easy to implement
lm_roll <- rollify(~lm(.x ~ .y), window = 5, unlist = FALSE)

gapminder %>%
  dplyr::group_by(country) %>%
  dplyr::mutate(rolling_lm = lm_roll(lifeExp, year))

#> [90m# A tibble: 1,704 x 7[39m
#> [90m# Groups:   country [142][39m
#>    country     continent  year lifeExp      pop gdpPercap rolling_lm
#>    [3m[90m<fct>[39m[23m       [3m[90m<fct>[39m[23m     [3m[90m<int>[39m[23m   [3m[90m<dbl>[39m[23m    [3m[90m<int>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<list>[39m[23m    
#> [90m 1[39m Afghanistan Asia       [4m1[24m952    28.8  8[4m4[24m[4m2[24m[4m5[24m333      779. [90m<lgl [1]>[39m 
#> [90m 2[39m Afghanistan Asia       [4m1[24m957    30.3  9[4m2[24m[4m4[24m[4m0[24m934      821. [90m<lgl [1]>[39m 
#> [90m 3[39m Afghanistan Asia       [4m1[24m962    32.0 10[4m2[24m[4m6[24m[4m7[24m083      853. [90m<lgl [1]>[39m 
#> [90m 4[39m Afghanistan Asia       [4m1[24m967    34.0 11[4m5[24m[4m3[24m[4m7[24m966      836. [90m<lgl [1]>[39m 
#> [90m 5[39m Afghanistan Asia       [4m1[24m972    36.1 13[4m0[24m[4m7[24m[4m9[24m460      740. [90m<S3: lm>[39m  
#> [90m 6[39m Afghanistan Asia       [4m1[24m977    38.4 14[4m8[24m[4m8[24m[4m0[24m372      786. [90m<S3: lm>[39m  
#> [90m 7[39m Afghanistan Asia       [4m1[24m982    39.9 12[4m8[24m[4m8[24m[4m1[24m816      978. [90m<S3: lm>[39m  
#> [90m 8[39m Afghanistan Asia       [4m1[24m987    40.8 13[4m8[24m[4m6[24m[4m7[24m957      852. [90m<S3: lm>[39m  
#> [90m 9[39m Afghanistan Asia       [4m1[24m992    41.7 16[4m3[24m[4m1[24m[4m7[24m921      649. [90m<S3: lm>[39m  
#> [90m10[39m Afghanistan Asia       [4m1[24m997    41.8 22[4m2[24m[4m2[24m[4m7[24m415      635. [90m<S3: lm>[39m  
#> [90m# … with 1,694 more rows[39m

# Rolling with groups -------------------------------------------------------

# One of the most powerful things about this is that it works with
# groups since `mutate` is being used
data(FANG)
FANG <- FANG %>%
  dplyr::group_by(symbol)

mean_roll_3 <- rollify(mean, window = 3)

FANG %>%
  dplyr::mutate(mean_roll = mean_roll_3(adjusted)) %>%
  dplyr::slice(1:5)

#> [90m# A tibble: 20 x 9[39m
#> [90m# Groups:   symbol [4][39m
#>    symbol date        open  high   low close   volume adjusted mean_roll
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m
#> [90m 1[39m AMZN   2013-01-02 256.  258.  253.  257.   3[4m2[24m[4m7[24m[4m1[24m000    257.       [31mNA[39m  
#> [90m 2[39m AMZN   2013-01-03 257.  261.  256.  258.   2[4m7[24m[4m5[24m[4m0[24m900    258.       [31mNA[39m  
#> [90m 3[39m AMZN   2013-01-04 258.  260.  257.  259.   1[4m8[24m[4m7[24m[4m4[24m200    259.      258. 
#> [90m 4[39m AMZN   2013-01-07 263.  270.  263.  268.   4[4m9[24m[4m1[24m[4m0[24m000    268.      262. 
#> [90m 5[39m AMZN   2013-01-08 267.  269.  264.  266.   3[4m0[24m[4m1[24m[4m0[24m700    266.      265. 
#> [90m 6[39m FB     2013-01-02  27.4  28.2  27.4  28   69[4m8[24m[4m4[24m[4m6[24m400     28        [31mNA[39m  
#> [90m 7[39m FB     2013-01-03  27.9  28.5  27.6  27.8 63[4m1[24m[4m4[24m[4m0[24m600     27.8      [31mNA[39m  
#> [90m 8[39m FB     2013-01-04  28.0  28.9  27.8  28.8 72[4m7[24m[4m1[24m[4m5[24m400     28.8      28.2
#> [90m 9[39m FB     2013-01-07  28.7  29.8  28.6  29.4 83[4m7[24m[4m8[24m[4m1[24m800     29.4      28.6
#> [90m10[39m FB     2013-01-08  29.5  29.6  28.9  29.1 45[4m8[24m[4m7[24m[4m1[24m300     29.1      29.1
#> [90m11[39m GOOG   2013-01-02 719.  727.  717.  723.   5[4m1[24m[4m0[24m[4m1[24m500    361.       [31mNA[39m  
#> [90m12[39m GOOG   2013-01-03 725.  732.  721.  724.   4[4m6[24m[4m5[24m[4m3[24m700    361.       [31mNA[39m  
#> [90m13[39m GOOG   2013-01-04 729.  741.  728.  738.   5[4m5[24m[4m4[24m[4m7[24m600    369.      364. 
#> [90m14[39m GOOG   2013-01-07 735.  739.  731.  735.   3[4m3[24m[4m2[24m[4m3[24m800    367.      366. 
#> [90m15[39m GOOG   2013-01-08 736.  736.  724.  733.   3[4m3[24m[4m6[24m[4m4[24m700    366.      367. 
#> [90m16[39m NFLX   2013-01-02  95.2  95.8  90.7  92.0 19[4m4[24m[4m3[24m[4m1[24m300     13.1      [31mNA[39m  
#> [90m17[39m NFLX   2013-01-03  92.0  97.9  91.5  96.6 27[4m9[24m[4m1[24m[4m2[24m500     13.8      [31mNA[39m  
#> [90m18[39m NFLX   2013-01-04  96.5  97.7  95.5  96.0 17[4m7[24m[4m6[24m[4m1[24m100     13.7      13.6
#> [90m19[39m NFLX   2013-01-07  96.4 102.   96.1  99.2 45[4m5[24m[4m5[24m[4m0[24m400     14.2      13.9
#> [90m20[39m NFLX   2013-01-08 100.  101.   96.8  97.2 24[4m7[24m[4m1[24m[4m4[24m900     13.9      13.9