tibbletime: time-aware tibbles

Built on top of the 'tibble' package, 'tibbletime' is an extension that allows for the creation of time aware tibbles. Some immediate advantages of this include: the ability to perform time based subsetting on tibbles, quickly summarising and aggregating results by time periods, and calling functions similar in spirit to the map family from 'purrr' on time based tibbles.

See also

Useful links:

https://github.com/business-science/tibbletime

Report bugs at https://github.com/business-science/tibbletime/issues