Collapse a tbl_time object so that all observations in a period share the same date

When time_collapse is used, the index of a tbl_time object is altered so that all dates that fall in a period share a common date. This can be useful for further groupwise calculations.

time_collapse(.data, period = "yearly", start_date = NULL,
  as_sep_col = FALSE, ...)

Arguments

.data
A tbl_time object.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

as_sep_col
Whether to keep the original index as the column .date or to drop it.

...
Not currently used.

Details

The date chosen as the common date for a period is always the date at the end of that period.

It is often useful to use as_sep_col = TRUE to keep the original dates as well.

This function respects dplyr::group_by() groups.

Examples

# Basic functionality -------------------------------------------------------

# Facebook stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Collapse to weekly dates
time_collapse(FB, period = "weekly")

#> # A time tibble: 1,008 x 8
#> # Index: date
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-04 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-04 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-11 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-11 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-11 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-11 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-18 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-18 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 998 more rows

# Collapse to every other week dates
time_collapse(FB, period = 2~w)

#> # A time tibble: 1,008 x 8
#> # Index: date
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-11 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-11 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-11 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-11 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-11 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-11 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-11 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-25 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-25 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 998 more rows

# Collapse to weekly dates, but keep the original too
time_collapse(FB, period = "weekly", as_sep_col = TRUE)

#> # A time tibble: 1,008 x 9
#> # Index: date
#>    symbol       date      .date  open  high   low close    volume adjusted
#>  *  <chr>     <date>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-04 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-04 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-04 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-11 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-11 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-11 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-11 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-11 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-18 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-18 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 998 more rows

# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- FANG %>%
  as_tbl_time(date) %>%
  dplyr::group_by(symbol)

# Collapse each group to monthly
FANG %>%
  time_collapse("monthly")

#> # A time tibble: 4,032 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-31 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-31 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-31 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-31 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-31 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-31 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-31 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-31 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-31 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-31 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 4,022 more rows

# Collapse each group to every other month
FANG %>%
  time_collapse(2~m)

#> # A time tibble: 4,032 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-02-28 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-02-28 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-02-28 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-02-28 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-02-28 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-02-28 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-02-28 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-02-28 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-02-28 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-02-28 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 4,022 more rows