Succinctly filter a tbl_time object by date

Use a concise filtering method to filter a tbl_time object by it's index.

time_filter(x, time_formula)

# S3 method for tbl_time
[(x, i, j, drop = FALSE)

Arguments

x
A tbl_time object.

time_formula
A period to filter over. This is specified as a formula. See Details.

i
A period to filter over. This is specified the same as time_formula or can use the traditional row extraction method.

j
Optional argument to also specify column index to subset. Works exactly like the normal extraction operator.

drop
Will always be coerced to FALSE by tibble.

Details

The time_formula is specified using the format from ~ to. Each side of the time_formula is specified as YYYY-MM-DD + HH:MM:SS, but powerful shorthand is available. Some examples are:

Year: 2013 ~ 2015

Month: 2013-01 ~ 2016-06

Day: 2013-01-05 ~ 2016-06-04

Second: 2013-01-05 + 10:22:15 ~ 2018-06-03 + 12:14:22

Variations: 2013 ~ 2016-06

The time_formula can also use a one sided formula.

Only dates in 2015: ~2015

Only dates March 2015: ~2015-03

All shorthand dates are expanded to be valid POSIXct dates.

The from side is expanded to be the first date in that period

The to side is expanded to be the last date in that period

This means that the following examples are equivalent:

2015 ~ 2016 == 2015-01-01 + 00:00:00 ~ 2016-12-31 + 23:59:59

~2015 == 2015-01-01 + 00:00:00 ~ 2015-12-31 + 23:59:59

2015-01-04 + 10:12 ~ 2015-01-05 == 2015-01-04 + 10:12:00 ~ 2015-01-05 + 23:59:59

This function respects dplyr::group_by() groups.

Examples

# FANG contains Facebook, Amazon, Netflix and Google stock prices
data(FANG)
FANG <- as_tbl_time(FANG, date) %>%
  dplyr::group_by(symbol)

# 2013-01-01 to 2014-12-31
time_filter(FANG, 2013 ~ 2014)

#> # A time tibble: 2,016 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 2,006 more rows

# 2013-05-25 to 2014-06-04
time_filter(FANG, 2013-05-25 ~ 2014-06-04)

#> # A time tibble: 1,032 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2013-05-28 24.54 24.54 23.92 24.10 50079700    24.10
#>  2     FB 2013-05-29 23.79 23.81 23.26 23.32 64237800    23.32
#>  3     FB 2013-05-30 24.13 24.78 23.93 24.55 60733200    24.55
#>  4     FB 2013-05-31 24.63 24.95 24.27 24.35 35925000    24.35
#>  5     FB 2013-06-03 24.27 24.32 23.71 23.85 35733800    23.85
#>  6     FB 2013-06-04 23.89 23.93 23.32 23.52 34760800    23.52
#>  7     FB 2013-06-05 23.35 23.71 22.79 22.90 53819700    22.90
#>  8     FB 2013-06-06 22.99 23.09 22.67 22.97 31260700    22.97
#>  9     FB 2013-06-07 23.03 23.40 22.86 23.29 38699200    23.29
#> 10     FB 2013-06-10 24.06 24.60 23.99 24.33 58393000    24.33
#> # ... with 1,022 more rows

# Using the `[` subset operator
FANG[2014~2015]

#> # A time tibble: 2,016 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2014-01-02 54.83 55.22 54.19 54.71 43195500    54.71
#>  2     FB 2014-01-03 55.02 55.65 54.53 54.56 38246200    54.56
#>  3     FB 2014-01-06 54.42 57.26 54.05 57.20 68852600    57.20
#>  4     FB 2014-01-07 57.70 58.55 57.22 57.92 77207400    57.92
#>  5     FB 2014-01-08 57.60 58.41 57.23 58.23 56682400    58.23
#>  6     FB 2014-01-09 58.65 58.96 56.65 57.22 92253300    57.22
#>  7     FB 2014-01-10 57.13 58.30 57.06 57.94 42449500    57.94
#>  8     FB 2014-01-13 57.91 58.25 55.38 55.91 63010900    55.91
#>  9     FB 2014-01-14 56.46 57.78 56.10 57.74 37503600    57.74
#> 10     FB 2014-01-15 57.98 58.57 57.27 57.60 33663400    57.60
#> # ... with 2,006 more rows

# Using `[` and one sided formula for only dates in 2014
FANG[~2014]

#> # A time tibble: 1,008 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2014-01-02 54.83 55.22 54.19 54.71 43195500    54.71
#>  2     FB 2014-01-03 55.02 55.65 54.53 54.56 38246200    54.56
#>  3     FB 2014-01-06 54.42 57.26 54.05 57.20 68852600    57.20
#>  4     FB 2014-01-07 57.70 58.55 57.22 57.92 77207400    57.92
#>  5     FB 2014-01-08 57.60 58.41 57.23 58.23 56682400    58.23
#>  6     FB 2014-01-09 58.65 58.96 56.65 57.22 92253300    57.22
#>  7     FB 2014-01-10 57.13 58.30 57.06 57.94 42449500    57.94
#>  8     FB 2014-01-13 57.91 58.25 55.38 55.91 63010900    55.91
#>  9     FB 2014-01-14 56.46 57.78 56.10 57.74 37503600    57.74
#> 10     FB 2014-01-15 57.98 58.57 57.27 57.60 33663400    57.60
#> # ... with 998 more rows

# Using `[` and column selection
FANG[2013~2016, c("date", "adjusted")]

#> # A time tibble: 4,032 x 2
#> # Index: date
#>          date adjusted
#>  *     <date>    <dbl>
#>  1 2013-01-02    28.00
#>  2 2013-01-03    27.77
#>  3 2013-01-04    28.76
#>  4 2013-01-07    29.42
#>  5 2013-01-08    29.06
#>  6 2013-01-09    30.59
#>  7 2013-01-10    31.30
#>  8 2013-01-11    31.72
#>  9 2013-01-14    30.95
#> 10 2013-01-15    30.10
#> # ... with 4,022 more rows

# You can filter with variables
lhs_date <- "2013"
rhs_date <- as.Date("2014-01-01")
time_filter(FANG, lhs_date ~ rhs_date)

#> Error in overscope_eval_next(overscope, expr): object 'lhs_date' not found