Floor a tbl_time object

This is essentially a convenient wrapper to lubridate::floor_date() that allows flooring of a tbl_time object using a period formula.

time_floor(x, period, ...)

Arguments

x
A tbl_time object.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

...
Arguments passed on to lubridate::floor_date()

Details

This function respects dplyr::group_by() groups.

Examples

example_series <- create_series(~2013, 2~day)

# When you convert to a period, the lowest available date in that period
# is chosen
example_series %>%
  as_period(1~m)

#> # A time tibble: 12 x 1
#> # Index: date
#>          date
#>  *     <date>
#>  1 2013-01-01
#>  2 2013-02-02
#>  3 2013-03-02
#>  4 2013-04-01
#>  5 2013-05-01
#>  6 2013-06-02
#>  7 2013-07-02
#>  8 2013-08-01
#>  9 2013-09-02
#> 10 2013-10-02
#> 11 2013-11-01
#> 12 2013-12-01

# Sometimes you want to additionally floor that date so that all dates
# are consistent
example_series %>%
  as_period(1~m) %>%
  time_floor(1~m)

#> # A time tibble: 12 x 1
#> # Index: date
#>          date
#>  *     <date>
#>  1 2013-01-01
#>  2 2013-02-01
#>  3 2013-03-01
#>  4 2013-04-01
#>  5 2013-05-01
#>  6 2013-06-01
#>  7 2013-07-01
#>  8 2013-08-01
#>  9 2013-09-01
#> 10 2013-10-01
#> 11 2013-11-01
#> 12 2013-12-01