Generate a vector of time-based groupings

time_group() accepts a date index vector and returns an integer vector that can be used for grouping by periods.

time_group(index, period = "yearly", start_date = NULL, ...)

Arguments

index
A vector of date indices to create groups for.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

...
Not currently used.

Details

This function is used internally, but may provide the user extra flexibility when they need to perform a grouped operation not supported by tibbletime.

Grouping can only be done on the minimum periodicity of the index and above. This means that a daily series cannot be grouped by minute. An hourly series cannot be grouped by 5 seconds, and so on. If the user attempts this, groups will be returned at the minimum periodicity (a daily series will return 1 group per day).

The start_date argument allows the user to control where the periods begin.

This function respects dplyr::group_by() groups.

See also

as_period(), create_series()

Examples

data(FB)

time_group(FB$date, 2~y)

#>    [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>   [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>   [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [149] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [297] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [334] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [371] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [408] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [445] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [482] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [519] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [556] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [593] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [630] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [667] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [704] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [741] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [778] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [815] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [852] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [889] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [926] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [963] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [1000] 2
#>  [ reached getOption("max.print") -- omitted 8 entries ]

dplyr::mutate(FB, time_group = time_group(date, 2~d))

#> # A tibble: 1,008 x 9
#>    symbol       date  open  high   low close    volume adjusted time_group
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>      <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00          1
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77          1
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76          2
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42          3
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06          4
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59          4
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30          5
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72          5
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95          7
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10          7
#> # ... with 998 more rows