Nest a tbl_time to a specified time period

time_nest allows for quick nesting of tbl_time objects by period so that manipulation can be performed on each period's data set individually. See tidyr::nest() for information on nesting generally.

time_nest(data, period = "yearly", ..., .key = "data")

Arguments

data
A tbl_time object.

period
The period to nest to.

...
Used to specify columns you do not want in the nest. Specified as -col_name.

.key
The name of the new column, as a string or symbol.

This argument is passed by expression and supports quasiquotation (you can unquote strings and symbols). The name is captured from the expression with rlang::quo_name() (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

Details

As an example, nesting by a "yearly" period will return a tbl_time object with a date column containing the dates at the end of each year, and a list-column of smaller tbl_time objects containing all of the data for that year.

Because it is likely the case that the original dates will be required inside the nested tbl_time objects for further manipulation, the original dates are kept as a .date column in each tbl_time object in the list-column.

This function respects dplyr::group_by() groups.

Note

The following periods are available:

"yearly"

"quarterly"

"monthly"

"weekly"

"daily"

"hour"

"minute"

"second"

Examples

# Basic functionality -------------------------------------------------------

data(FB)
FB <- as_tbl_time(FB, date)

# Nest yearly
time_nest(FB, "yearly")

#> # A time tibble: 4 x 2
#> # Index: date
#>         date               data
#> *     <date>             <list>
#> 1 2013-12-31 <tibble [252 x 8]>
#> 2 2014-12-31 <tibble [252 x 8]>
#> 3 2015-12-31 <tibble [252 x 8]>
#> 4 2016-12-30 <tibble [252 x 8]>

# Nest yearly, but don't drag symbol into the nest
time_nest(FB, "yearly", -symbol)

#> # A time tibble: 4 x 3
#> # Index: date
#>   symbol       date               data
#> *  <chr>     <date>             <list>
#> 1     FB 2013-12-31 <tibble [252 x 7]>
#> 2     FB 2014-12-31 <tibble [252 x 7]>
#> 3     FB 2015-12-31 <tibble [252 x 7]>
#> 4     FB 2016-12-30 <tibble [252 x 7]>

# Nest quarterly
time_nest(FB, "quarterly")

#> # A time tibble: 16 x 2
#> # Index: date
#>          date              data
#>  *     <date>            <list>
#>  1 2013-03-28 <tibble [60 x 8]>
#>  2 2013-06-28 <tibble [64 x 8]>
#>  3 2013-09-30 <tibble [64 x 8]>
#>  4 2013-12-31 <tibble [64 x 8]>
#>  5 2014-03-31 <tibble [61 x 8]>
#>  6 2014-06-30 <tibble [63 x 8]>
#>  7 2014-09-30 <tibble [64 x 8]>
#>  8 2014-12-31 <tibble [64 x 8]>
#>  9 2015-03-31 <tibble [61 x 8]>
#> 10 2015-06-30 <tibble [63 x 8]>
#> 11 2015-09-30 <tibble [64 x 8]>
#> 12 2015-12-31 <tibble [64 x 8]>
#> 13 2016-03-31 <tibble [61 x 8]>
#> 14 2016-06-30 <tibble [64 x 8]>
#> 15 2016-09-30 <tibble [64 x 8]>
#> 16 2016-12-30 <tibble [63 x 8]>

# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- as_tbl_time(FANG, date) %>%
  group_by(symbol)

# Nest yearly, but by group
FANG %>%
  time_nest("yearly")

#> # A time tibble: 16 x 3
#> # Index: date
#>    symbol       date               data
#>  *  <chr>     <date>             <list>
#>  1     FB 2013-12-31 <tibble [252 x 7]>
#>  2     FB 2014-12-31 <tibble [252 x 7]>
#>  3     FB 2015-12-31 <tibble [252 x 7]>
#>  4     FB 2016-12-30 <tibble [252 x 7]>
#>  5   AMZN 2013-12-31 <tibble [252 x 7]>
#>  6   AMZN 2014-12-31 <tibble [252 x 7]>
#>  7   AMZN 2015-12-31 <tibble [252 x 7]>
#>  8   AMZN 2016-12-30 <tibble [252 x 7]>
#>  9   NFLX 2013-12-31 <tibble [252 x 7]>
#> 10   NFLX 2014-12-31 <tibble [252 x 7]>
#> 11   NFLX 2015-12-31 <tibble [252 x 7]>
#> 12   NFLX 2016-12-30 <tibble [252 x 7]>
#> 13   GOOG 2013-12-31 <tibble [252 x 7]>
#> 14   GOOG 2014-12-31 <tibble [252 x 7]>
#> 15   GOOG 2015-12-31 <tibble [252 x 7]>
#> 16   GOOG 2016-12-30 <tibble [252 x 7]>