Summarise a tbl_time by period

time_summarise() works similarly to dplyr::summarise() but with the added benefit of being able to summarise by a time period such as "yearly" or "monthly".

time_summarise(.data, period = "yearly", ..., start_date = NULL)

time_summarize(.data, period = "yearly", ..., start_date = NULL)

Arguments

.data
A tbl_time object.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

...
Not currently used.

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

Details

Groups applied using dplyr::group_by() are respected.

In dplyr::summarise(), one level of grouping is usually removed. Because an added group for the time index is added in time_summarise, none of the original groups are removed.

Examples

# Basic functionality -------------------------------------------------------

data(FB)
FB <- as_tbl_time(FB, date)

# Calculate the mean and standard deviation of the adjusted column
# at a yearly interval
FB %>%
  time_summarise(period = "yearly",
                 adj_mean = mean(adjusted),
                 adj_sd   = sd(adjusted))

#> # A time tibble: 4 x 3
#> # Index: date
#>         date  adj_mean    adj_sd
#> *     <date>     <dbl>     <dbl>
#> 1 2013-12-31  35.48115 10.621172
#> 2 2014-12-31  68.76234  7.502259
#> 3 2015-12-31  88.77286 10.211442
#> 4 2016-12-30 117.03587  8.899858

# Want a more granular view? Look at monthly instead
FB %>%
  time_summarise(period = "monthly",
                 adj_mean = mean(adjusted),
                 adj_sd   = sd(adjusted))

#> # A time tibble: 48 x 3
#> # Index: date
#>          date adj_mean    adj_sd
#>  *     <date>    <dbl>     <dbl>
#>  1 2013-01-31 30.33190 1.2134688
#>  2 2013-02-28 28.08789 0.7823224
#>  3 2013-03-28 26.80650 1.0487099
#>  4 2013-04-30 26.60818 0.7453578
#>  5 2013-05-31 26.14318 1.4524016
#>  6 2013-06-28 23.94400 0.5166124
#>  7 2013-07-31 27.91136 4.3982598
#>  8 2013-08-30 38.74818 1.4761225
#>  9 2013-09-30 45.76200 2.9370909
#> 10 2013-10-31 50.51174 1.8817587
#> # ... with 38 more rows

# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- as_tbl_time(FANG, date) %>%
  dplyr::group_by(symbol)

# Groups are respected, allowing for very useful summaries
# grouped by symbol and by time period
FANG %>%
  time_summarise(period = "yearly",
                 vol_max   = max(volume),
                 vol_min   = min(volume),
                 # Like summarise, you can use columns you just computed
                 vol_range = vol_max - vol_min)

#> # A time tibble: 16 x 5
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date   vol_max  vol_min vol_range
#>  *  <chr>     <date>     <dbl>    <dbl>     <dbl>
#>  1   AMZN 2013-12-31  14030000  1350600  12679400
#>  2   AMZN 2014-12-31  19801100  1513800  18287300
#>  3   AMZN 2015-12-31  23856100  1091200  22764900
#>  4   AMZN 2016-12-30  14677600  1458800  13218800
#>  5     FB 2013-12-31 365457900 10404400 355053500
#>  6     FB 2014-12-31 150178900  7685500 142493400
#>  7     FB 2015-12-31  86211300  5913100  80298200
#>  8     FB 2016-12-30 107475300  8638400  98836900
#>  9   GOOG 2013-12-31  23219400  1473800  21745600
#> 10   GOOG 2014-12-31  11174200     7900  11166300
#> 11   GOOG 2015-12-31  11164900   527200  10637700
#> 12   GOOG 2016-12-30   6348100   586000   5762100
#> 13   NFLX 2013-12-31 191445800  5577600 185868200
#> 14   NFLX 2014-12-31  92304800  5411000  86893800
#> 15   NFLX 2015-12-31 104500900  3531300 100969600
#> 16   NFLX 2016-12-30  55728800  1616300  54112500