Apply a function to each element of a tbl_time object.

The tmap functions transform a tbl_time input by applying a function to each column at a specified time interval.

tmap(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_chr(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_int(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_lgl(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_dbl(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_dfc(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_dfr(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

tmap_df(.x, .f, period = "yearly", start_date = NULL, name = "data", ...)

Arguments

.x
A tbl_time object.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get_attr() to extract named attributes. If a component is not present, the value of .default will be returned.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

name
The character name of the list-column generated.

...
Not currently used.

Value

A tbl_time object grouped by the time interval specified. The last available date in that interval is returned as the new date.

Details

These functions are similar to purrr::map(), but allow the user to also perform grouped mapping over intervals such as "yearly", "monthly", "hourly", etc. For example, if "monthly" is chosen, one could picture the tbl_time object being split into smaller tbl_times, one for each month, and having the function mapped over all of the columns in each of those smaller tibbles. The results are then recombined into one tibble, with a list-column holding the results of the mapping over each time period.

Groupings applied using dplyr::group_by() are respected.

Examples

# First example -------------------------------------------------------------

data(FB)
FB <- as_tbl_time(FB, date)
# No need for the symbol column here
FB <- dplyr::select(FB, -symbol)

# Get the yearly average of every column in FB
mapped_mean <- FB %>%
  tmap(.f = ~mean(.x), period = "yearly")

# It is returned as a list-column because the time period adds
# an extra dimension to the mapping
mapped_mean

#> # A time tibble: 4 x 2
#> # Index: date
#>         date       data
#> *     <date>     <list>
#> 1 2013-12-31 <list [6]>
#> 2 2014-12-31 <list [6]>
#> 3 2015-12-31 <list [6]>
#> 4 2016-12-30 <list [6]>

# Access individual elements. Here, the 2013 results
mapped_mean$data[[1]]

#> $open
#> [1] 35.47643
#> 
#> $high
#> [1] 36.04504
#> 
#> $low
#> [1] 34.93016
#> 
#> $close
#> [1] 35.48115
#> 
#> $volume
#> [1] 60091994
#> 
#> $adjusted
#> [1] 35.48115
#> 

# More useful example -------------------------------------------------------

data(FB)
FB <- as_tbl_time(FB, date)
# No need for the symbol column here
FB <- dplyr::select(FB, -symbol)

#> Error in overscope_eval_next(overscope, expr): object 'symbol' not found

# An easier approach might be to use `tmap_dfc` to coerce each list-column
# entry to a tibble, then unnest the result
# Here we calculate the monthly average for each column
FB %>%
  tmap_dfc(~mean(.x), period = "monthly") %>%
  tidyr::unnest()

#> # A time tibble: 48 x 7
#> # Index: date
#>          date     open     high      low    close   volume adjusted
#>  *     <date>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2013-01-31 30.18762 30.83143 29.77000 30.33190 79802462 30.33190
#>  2 2013-02-28 28.26316 28.59895 27.73211 28.08789 50402095 28.08789
#>  3 2013-03-28 26.88050 27.24850 26.45300 26.80650 36359025 26.80650
#>  4 2013-04-30 26.58864 27.01818 26.21864 26.60818 33568600 26.60818
#>  5 2013-05-31 26.35136 26.63273 25.91455 26.14318 44640673 26.14318
#>  6 2013-06-28 23.98250 24.29850 23.65600 23.94400 39416575 23.94400
#>  7 2013-07-31 27.72818 28.21818 27.39000 27.91136 65364414 27.91136
#>  8 2013-08-30 38.70045 39.28500 38.23591 38.74818 61136095 38.74818
#>  9 2013-09-30 45.53850 46.31400 44.90200 45.76200 79154190 45.76200
#> 10 2013-10-31 50.67348 51.51348 49.71087 50.51174 88375435 50.51174
#> # ... with 38 more rows

# Functions with multiple return values -------------------------------------

data(FB)
FB <- as_tbl_time(FB, date)
# No need for the symbol column here
FB <- dplyr::select(FB, -symbol)

#> Error in overscope_eval_next(overscope, expr): object 'symbol' not found

# Functions that return more than 1 number per map are possible, but more
# difficult to work with.
# Mapping the quantile function to each column of FB at yearly time increments
mapped_quantile <- FB %>%
  tmap(~quantile(.x), "yearly")

mapped_quantile$data[[1]]

#> $open
#>      0%     25%     50%     75%    100% 
#> 22.9900 26.5675 29.7100 46.4975 58.3200 
#> 
#> $high
#>      0%     25%     50%     75%    100% 
#> 23.0900 26.7875 30.4300 47.3925 58.5800 
#> 
#> $low
#>      0%     25%     50%     75%    100% 
#> 22.6700 26.1175 29.3800 45.8075 57.3700 
#> 
#> $close
#>     0%    25%    50%    75%   100% 
#> 22.900 26.505 29.790 46.520 57.960 
#> 
#> $volume
#>        0%       25%       50%       75%      100% 
#>  10404400  34321200  49838850  75650175 365457900 
#> 
#> $adjusted
#>     0%    25%    50%    75%   100% 
#> 22.900 26.505 29.790 46.520 57.960 
#> 

# It is possible to get a cleaner result, but currently not intuitive
library(tibble)
FB %>%
  tidyr::gather(key = col_type, value = value, -date) %>%
  dplyr::group_by(col_type) %>%
  tmap_dfr(~quantile(.x) %>% as.list %>% as.tibble, period = "yearly") %>%
  tidyr::unnest()

#> # A time tibble: 24 x 7
#> # Index:  date
#> # Groups: col_type [6]
#>          date col_type  `0%`    `25%`   `50%`    `75%` `100%`
#>  *     <date>    <chr> <dbl>    <dbl>   <dbl>    <dbl>  <dbl>
#>  1 2013-12-31     open 22.99  26.5675  29.710  46.4975  58.32
#>  2 2014-12-31     open 54.02  62.6025  69.255  75.1775  82.02
#>  3 2015-12-31     open 74.04  79.9575  86.750  95.4550 108.80
#>  4 2016-12-30     open 92.83 111.7650 117.845 123.7075 133.50
#>  5 2013-12-31     high 23.09  26.7875  30.430  47.3925  58.58
#>  6 2014-12-31     high 54.94  63.4350  69.960  75.8725  82.17
#>  7 2015-12-31     high 74.83  81.0925  87.760  96.3700 110.65
#>  8 2016-12-30     high 95.00 112.8175 118.680 124.6300 133.50
#>  9 2013-12-31      low 22.67  26.1175  29.380  45.8075  57.37
#> 10 2014-12-31      low 51.85  61.4125  68.375  74.5300  80.75
#> # ... with 14 more rows