Change tbl_time periodicity

Convert a tbl_time object from daily to monthly, from minute data to hourly, and more. This allows the user to easily aggregate data to a less granular level.

as_period(x, period = "yearly", side = "start", start_date = NULL)

Arguments

x
A tbl_time object.

period
A formula or character specification used for time-based grouping.

If a formula, e.g. 1~year, the formula is split and parsed to form the grouping period. The period argument accepts a formula of the form multiple ~ period allowing for flexible period grouping. The following are examples:

1 Year: 1~y

3 Months: 3~m

90 Days: 90~d

Note that while shorthand is used above, an attempt is made to recognize more explicit period names such as:

2 Year: 2~year / 2~years / 2~yearly

The period argument also accepts characters that are converted to their corresponding periods. The following are accepted:

"yearly" or "y"

"quarterly" or "q"

"monthly" or "m"

"weekly" or "w"

"daily" or "d"

"hour" or "h"

"minute" or "M"

"second" or "s"

side
Whether to return the date at the beginning or the end of the new period. By default, the "start" of the period. Use "end" to change to the end of the period.

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

Details

This function respects dplyr::group_by() groups.

Currently periods finer than second data are not supported.

The side argument is useful when you want to return data at, say, the end of a quarter, or the end of a month.

Examples

# Basic usage ---------------------------------------------------------------

# FB stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Aggregate FB to yearly data
as_period(FB, "yearly")

#> # A time tibble: 4 x 8
#> # Index: date
#>   symbol       date   open   high   low  close   volume adjusted
#> *  <chr>     <date>  <dbl>  <dbl> <dbl>  <dbl>    <dbl>    <dbl>
#> 1     FB 2013-01-02  27.44  28.18 27.42  28.00 69846400    28.00
#> 2     FB 2014-01-02  54.83  55.22 54.19  54.71 43195500    54.71
#> 3     FB 2015-01-02  78.58  78.93 77.70  78.45 18177500    78.45
#> 4     FB 2016-01-04 101.95 102.24 99.75 102.22 37912400   102.22

# Aggregate FB to yearly data using a period formula
as_period(FB, 1~y)

#> # A time tibble: 4 x 8
#> # Index: date
#>   symbol       date   open   high   low  close   volume adjusted
#> *  <chr>     <date>  <dbl>  <dbl> <dbl>  <dbl>    <dbl>    <dbl>
#> 1     FB 2013-01-02  27.44  28.18 27.42  28.00 69846400    28.00
#> 2     FB 2014-01-02  54.83  55.22 54.19  54.71 43195500    54.71
#> 3     FB 2015-01-02  78.58  78.93 77.70  78.45 18177500    78.45
#> 4     FB 2016-01-04 101.95 102.24 99.75 102.22 37912400   102.22

# Aggregate FB to yearly data, but use the last data point available
# in that period
as_period(FB, "yearly", "end")

#> # A time tibble: 4 x 8
#> # Index: date
#>   symbol       date   open   high    low  close   volume adjusted
#> *  <chr>     <date>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>    <dbl>
#> 1     FB 2013-12-31  54.12  54.86  53.91  54.65 43076200    54.65
#> 2     FB 2014-12-31  79.54  79.80  77.86  78.02 19935400    78.02
#> 3     FB 2015-12-31 106.00 106.17 104.62 104.66 18298700   104.66
#> 4     FB 2016-12-30 116.60 116.83 114.77 115.05 18600100   115.05

# Aggregate to weekly. Notice that it only uses the earliest day available
# in the data set at that periodicity. It will not set the date of the first
# row to 2013-01-01 because that date did not exist in the original data set.
as_period(FB, "weekly")

#> # A time tibble: 209 x 8
#> # Index: date
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00 69846400    28.00
#>  2     FB 2013-01-07 28.69 29.79 28.65 29.42 83781800    29.42
#>  3     FB 2013-01-14 32.08 32.21 30.62 30.95 98892800    30.95
#>  4     FB 2013-01-22 29.75 30.89 29.74 30.73 55243300    30.73
#>  5     FB 2013-01-28 31.88 32.51 31.81 32.47 59682500    32.47
#>  6     FB 2013-02-04 29.06 29.20 28.01 28.11 92362200    28.11
#>  7     FB 2013-02-11 28.61 28.68 28.04 28.26 37361800    28.26
#>  8     FB 2013-02-19 28.23 29.08 28.12 28.93 49396400    28.93
#>  9     FB 2013-02-25 27.16 27.64 27.15 27.27 34652000    27.27
#> 10     FB 2013-03-04 27.76 28.06 27.44 27.72 32400700    27.72
#> # ... with 199 more rows

# Aggregate to every other week
as_period(FB, 2~w)

#> # A time tibble: 105 x 8
#> # Index: date
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00 69846400    28.00
#>  2     FB 2013-01-14 32.08 32.21 30.62 30.95 98892800    30.95
#>  3     FB 2013-01-28 31.88 32.51 31.81 32.47 59682500    32.47
#>  4     FB 2013-02-11 28.61 28.68 28.04 28.26 37361800    28.26
#>  5     FB 2013-02-25 27.16 27.64 27.15 27.27 34652000    27.27
#>  6     FB 2013-03-11 28.01 28.64 27.83 28.14 35642100    28.14
#>  7     FB 2013-03-25 25.75 25.80 25.08 25.13 39199000    25.13
#>  8     FB 2013-04-08 27.19 27.20 26.63 26.85 27256000    26.85
#>  9     FB 2013-04-22 25.81 26.36 25.70 25.97 25687600    25.97
#> 10     FB 2013-05-06 28.33 28.46 27.48 27.57 43939400    27.57
#> # ... with 95 more rows

# FB is daily data, aggregate to minute?
# Does nothing and returns the original data set.
as_period(FB, "minute")

#> # A time tibble: 1,008 x 8
#> # Index: date
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10
#> # ... with 998 more rows

# Grouped usage -------------------------------------------------------------

# FANG contains Facebook, Amazon, Netflix and Google stock prices
data(FANG)
FANG <- as_tbl_time(FANG, date)

FANG <- FANG %>% group_by(symbol)

# Respects groups
FANG %>% as_period("yearly")

#> # A time tibble: 16 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date      open      high       low     close   volume  adjusted
#>  *  <chr>     <date>     <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1     FB 2013-01-02   27.4400   28.1800   27.4200   28.0000 69846400  28.00000
#>  2     FB 2014-01-02   54.8300   55.2200   54.1900   54.7100 43195500  54.71000
#>  3     FB 2015-01-02   78.5800   78.9300   77.7000   78.4500 18177500  78.45000
#>  4     FB 2016-01-04  101.9500  102.2400   99.7500  102.2200 37912400 102.22000
#>  5   AMZN 2013-01-02  256.0800  258.1000  253.2600  257.3100  3271000 257.31000
#>  6   AMZN 2014-01-02  398.8000  399.3600  394.0200  397.9700  2137800 397.97000
#>  7   AMZN 2015-01-02  312.5800  314.7500  306.9600  308.5200  2783200 308.51999
#>  8   AMZN 2016-01-04  656.2900  657.7200  627.5100  636.9900  9314500 636.98999
#>  9   NFLX 2013-01-02   95.2100   95.8100   90.6900   92.0100 19431300  13.14429
#> 10   NFLX 2014-01-02  366.8100  367.5800  360.8000  362.8200 12325600  51.83143
#> 11   NFLX 2015-01-02  344.0600  352.3200  341.1200  348.9400 13475000  49.84857
#> 12   NFLX 2016-01-04  109.0000  110.0000  105.2100  109.9600 20794800 109.96000
#> 13   GOOG 2013-01-02  719.4212  727.0013  716.5512  723.2512  5101500 361.26435
#> 14   GOOG 2014-01-02 1115.4619 1117.7519 1108.2619 1113.1220  3656400 556.00497
#> 15   GOOG 2015-01-02  529.0124  531.2724  524.1024  524.8124  1447500 524.81240
#> 16   GOOG 2016-01-04  743.0000  744.0600  731.2580  741.8400  3272800 741.84003

# Every 6 months, respecting groups
as_period(FANG, 6~m)

#> # A time tibble: 32 x 8
#> # Index:  date
#> # Groups: symbol [4]
#>    symbol       date   open   high    low  close   volume adjusted
#>  *  <chr>     <date>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>    <dbl>
#>  1     FB 2013-01-02  27.44  28.18  27.42  28.00 69846400    28.00
#>  2     FB 2013-07-01  24.97  25.06  24.62  24.81 20582200    24.81
#>  3     FB 2014-01-02  54.83  55.22  54.19  54.71 43195500    54.71
#>  4     FB 2014-07-01  67.58  68.44  67.39  68.06 33243000    68.06
#>  5     FB 2015-01-02  78.58  78.93  77.70  78.45 18177500    78.45
#>  6     FB 2015-07-01  86.77  87.95  86.49  86.91 25260000    86.91
#>  7     FB 2016-01-04 101.95 102.24  99.75 102.22 37912400   102.22
#>  8     FB 2016-07-01 114.20 115.13 113.81 114.19 14980000   114.19
#>  9   AMZN 2013-01-02 256.08 258.10 253.26 257.31  3271000   257.31
#> 10   AMZN 2013-07-01 279.00 283.29 277.16 282.10  2888200   282.10
#> # ... with 22 more rows

# Using start_date ----------------------------------------------------------

# FB stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# The Facebook series starts at 2013-01-02 so the 'every 2 day' counter
# starts at that date as well. Groups become (2013-01-02, 2013-01-03),
# (2013-01-04, 2013-01-05) and so on.
as_period(FB, 2~d)

#> # A time tibble: 607 x 8
#> # Index: date
#>    symbol       date  open  high   low close   volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00 69846400    28.00
#>  2     FB 2013-01-04 28.01 28.93 27.83 28.76 72715400    28.76
#>  3     FB 2013-01-07 28.69 29.79 28.65 29.42 83781800    29.42
#>  4     FB 2013-01-08 29.51 29.60 28.86 29.06 45871300    29.06
#>  5     FB 2013-01-10 30.60 31.45 30.28 31.30 95316400    31.30
#>  6     FB 2013-01-14 32.08 32.21 30.62 30.95 98892800    30.95
#>  7     FB 2013-01-16 30.21 30.35 29.53 29.85 75332700    29.85
#>  8     FB 2013-01-18 30.31 30.44 29.27 29.66 49631500    29.66
#>  9     FB 2013-01-22 29.75 30.89 29.74 30.73 55243300    30.73
#> 10     FB 2013-01-24 31.27 31.49 30.81 31.08 43845100    31.08
#> # ... with 597 more rows

# Specifying the `start_date = "2013-01-01"` might be preferable.
# Groups become (2013-01-01, 2013-01-02), (2013-01-03, 2013-01-04) and so on.
as_period(FB, 2~d, start_date = "2013-01-01")

#> # A time tibble: 619 x 8
#> # Index: date
#>    symbol       date  open  high   low close    volume adjusted
#>  *  <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77
#>  3     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42
#>  4     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59
#>  5     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72
#>  6     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95
#>  7     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10
#>  8     FB 2013-01-17 30.08 30.42 30.03 30.14  40256700    30.14
#>  9     FB 2013-01-22 29.75 30.89 29.74 30.73  55243300    30.73
#> 10     FB 2013-01-23 31.10 31.50 30.80 30.82  48899800    30.82
#> # ... with 609 more rows