Collapse a tbl_time object by its index

Collapse the index of a tbl_time object by time period. The index column is altered so that all dates that fall in a specified interval share a common date.

collapse_by(.tbl_time, period = "yearly", start_date = NULL,
  side = "end", clean = FALSE, ...)

Arguments

.tbl_time
A tbl_time object.

period
A character specification used for time-based grouping. The general format to use is "frequency period" where frequency is a number like 1 or 2, and period is an interval like weekly or yearly. There must be a space between the two.

Note that you can pass the specification in a flexible way:

1 Year: '1 year' / '1 Y' / '1 yearly' / 'yearly'

This shorthand is available for year, quarter, month, day, hour, minute, second, millisecond and microsecond periodicities.

Additionally, you have the option of passing in a vector of dates to use as custom and more flexible boundaries.

start_date
Optional argument used to specify the start date for the first group. The default is to start at the closest period boundary below the minimum date in the supplied index.

side
Whether to return the date at the beginning or the end of the new period. By default, the "end" of the period. Use "start" to change to the start of the period.

clean
Whether or not to round the collapsed index up / down to the next period boundary. The decision to round up / down is controlled by the side argument.

...
Not currently used.

Details

collapse_by() is a simplification of a call to dplyr::mutate() to collapse an index column using collapse_index().

Examples

# Basic functionality -------------------------------------------------------

# Facebook stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Collapse to weekly dates
collapse_by(FB, "weekly")

#> # A time tibble: 1,008 x 8
#> [90m# Index: date[39m
#>    symbol date        open  high   low close    volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m FB     2013-01-04  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28  
#> [90m 2[39m FB     2013-01-04  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8  72[4m7[24m[4m1[24m[4m5[24m400     28.8
#> [90m 4[39m FB     2013-01-11  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4
#> [90m 5[39m FB     2013-01-11  29.5  29.6  28.9  29.1  45[4m8[24m[4m7[24m[4m1[24m300     29.1
#> [90m 6[39m FB     2013-01-11  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6
#> [90m 7[39m FB     2013-01-11  30.6  31.5  30.3  31.3  95[4m3[24m[4m1[24m[4m6[24m400     31.3
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7
#> [90m 9[39m FB     2013-01-18  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0
#> [90m10[39m FB     2013-01-18  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1
#> [90m# … with 998 more rows[39m

# A common workflow is to group on the collapsed date column
# to perform a time based summary
FB %>%
  collapse_by("yearly") %>%
  dplyr::group_by(date) %>%
  dplyr::summarise_if(is.numeric, mean)

#> # A time tibble: 4 x 7
#> [90m# Index: date[39m
#>   date        open  high   low close    volume adjusted
#>   [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m 2013-12-31  35.5  36.0  34.9  35.5 60[4m0[24m[4m9[24m[4m1[24m994.     35.5
#> [90m2[39m 2014-12-31  68.8  69.6  67.8  68.8 47[4m5[24m[4m3[24m[4m0[24m552.     68.8
#> [90m3[39m 2015-12-31  88.7  89.7  87.8  88.8 26[4m9[24m[4m5[24m[4m5[24m191.     88.8
#> [90m4[39m 2016-12-30 117.  118.  116.  117.  25[4m4[24m[4m5[24m[4m3[24m798.    117. 

# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- FANG %>%
  as_tbl_time(date) %>%
  dplyr::group_by(symbol)

# Collapse each group to monthly,
# calculate monthly standard deviation for each column
FANG %>%
  collapse_by("monthly") %>%
  dplyr::group_by(date, add = TRUE) %>%
  dplyr::summarise_all(sd)

#> # A time tibble: 192 x 8
#> [90m# Index:  date[39m
#> [90m# Groups: symbol [?][39m
#>    symbol date        open  high   low close   volume adjusted
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m AMZN   2013-01-31  7.11  7.21  5.63  6.31 2[4m8[24m[4m5[24m[4m1[24m878.     6.31
#> [90m 2[39m AMZN   2013-02-28  3.21  3.38  3.26  3.82  [4m8[24m[4m6[24m[4m3[24m971.     3.82
#> [90m 3[39m AMZN   2013-03-28  7.50  7.29  7.51  7.67  [4m9[24m[4m2[24m[4m9[24m111.     7.67
#> [90m 4[39m AMZN   2013-04-30  6.32  6.22  6.15  6.69 2[4m6[24m[4m8[24m[4m8[24m623.     6.69
#> [90m 5[39m AMZN   2013-05-31  5.40  5.60  6.44  5.72  [4m7[24m[4m8[24m[4m1[24m265.     5.72
#> [90m 6[39m AMZN   2013-06-28  4.29  4.45  4.91  4.70  [4m6[24m[4m9[24m[4m2[24m542.     4.70
#> [90m 7[39m AMZN   2013-07-31 10.1   9.68  9.09  9.04 1[4m5[24m[4m9[24m[4m2[24m530.     9.04
#> [90m 8[39m AMZN   2013-08-30  7.67  7.08  6.98  7.65  [4m4[24m[4m4[24m[4m3[24m340.     7.65
#> [90m 9[39m AMZN   2013-09-30  9.67  9.37  9.37  9.18  [4m8[24m[4m4[24m[4m7[24m804.     9.18
#> [90m10[39m AMZN   2013-10-31 20.6  21.4  20.7  21.4  2[4m2[24m[4m0[24m[4m9[24m026.    21.4 
#> [90m# … with 182 more rows[39m