Create a tbl_time object with a sequence of regularly spaced dates

create_series() allows the user to quickly create a tbl_time object with a date column populated with a sequence of dates.

create_series(time_formula, period = "daily", class = "POSIXct",
  include_end = FALSE, tz = "UTC", as_vector = FALSE)

Arguments

time_formula
A period to create the series over. This is specified as a formula. See the Details section of filter_time() for more information.

period
A character specification used for time-based grouping. The general format to use is "frequency period" where frequency is a number like 1 or 2, and period is an interval like weekly or yearly. There must be a space between the two.

Note that you can pass the specification in a flexible way:

1 Year: '1 year' / '1 Y' / '1 yearly' / 'yearly'

This shorthand is available for year, quarter, month, day, hour, minute, second, millisecond and microsecond periodicities.

Additionally, you have the option of passing in a vector of dates to use as custom and more flexible boundaries.

class
One of "Date", "POSIXct", "hms", "yearmon", "yearqtr". The default is "POSIXct".

include_end
Whether to always include the RHS of the time_formula even if it does not match the regularly spaced index.

tz
Time zone of the new series.

as_vector
Should the series be returned as a vector instead of a tibble?

Examples

# Every day in 2013
create_series(~'2013', 'daily')

#> # A time tibble: 365 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2013-01-01 [90m00:00:00[39m
#> [90m 2[39m 2013-01-02 [90m00:00:00[39m
#> [90m 3[39m 2013-01-03 [90m00:00:00[39m
#> [90m 4[39m 2013-01-04 [90m00:00:00[39m
#> [90m 5[39m 2013-01-05 [90m00:00:00[39m
#> [90m 6[39m 2013-01-06 [90m00:00:00[39m
#> [90m 7[39m 2013-01-07 [90m00:00:00[39m
#> [90m 8[39m 2013-01-08 [90m00:00:00[39m
#> [90m 9[39m 2013-01-09 [90m00:00:00[39m
#> [90m10[39m 2013-01-10 [90m00:00:00[39m
#> [90m# … with 355 more rows[39m

# Every other day in 2013
create_series(~'2013', '2 d')

#> # A time tibble: 183 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2013-01-01 [90m00:00:00[39m
#> [90m 2[39m 2013-01-03 [90m00:00:00[39m
#> [90m 3[39m 2013-01-05 [90m00:00:00[39m
#> [90m 4[39m 2013-01-07 [90m00:00:00[39m
#> [90m 5[39m 2013-01-09 [90m00:00:00[39m
#> [90m 6[39m 2013-01-11 [90m00:00:00[39m
#> [90m 7[39m 2013-01-13 [90m00:00:00[39m
#> [90m 8[39m 2013-01-15 [90m00:00:00[39m
#> [90m 9[39m 2013-01-17 [90m00:00:00[39m
#> [90m10[39m 2013-01-19 [90m00:00:00[39m
#> [90m# … with 173 more rows[39m

# Every quarter in 2013
create_series(~'2013', '1 q')

#> # A time tibble: 4 x 1
#> [90m# Index: date[39m
#>   date               
#>   [3m[90m<dttm>[39m[23m             
#> [90m1[39m 2013-01-01 [90m00:00:00[39m
#> [90m2[39m 2013-04-01 [90m00:00:00[39m
#> [90m3[39m 2013-07-01 [90m00:00:00[39m
#> [90m4[39m 2013-10-01 [90m00:00:00[39m

# Daily series for 2013-2015
create_series('2013' ~ '2015', '1 d')

#> # A time tibble: 1,095 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2013-01-01 [90m00:00:00[39m
#> [90m 2[39m 2013-01-02 [90m00:00:00[39m
#> [90m 3[39m 2013-01-03 [90m00:00:00[39m
#> [90m 4[39m 2013-01-04 [90m00:00:00[39m
#> [90m 5[39m 2013-01-05 [90m00:00:00[39m
#> [90m 6[39m 2013-01-06 [90m00:00:00[39m
#> [90m 7[39m 2013-01-07 [90m00:00:00[39m
#> [90m 8[39m 2013-01-08 [90m00:00:00[39m
#> [90m 9[39m 2013-01-09 [90m00:00:00[39m
#> [90m10[39m 2013-01-10 [90m00:00:00[39m
#> [90m# … with 1,085 more rows[39m

# Minute series for 2 months
create_series('2012-01' ~ '2012-02', 'M')

#> # A time tibble: 86,400 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2012-01-01 [90m00:00:00[39m
#> [90m 2[39m 2012-01-01 [90m00:01:00[39m
#> [90m 3[39m 2012-01-01 [90m00:02:00[39m
#> [90m 4[39m 2012-01-01 [90m00:03:00[39m
#> [90m 5[39m 2012-01-01 [90m00:04:00[39m
#> [90m 6[39m 2012-01-01 [90m00:05:00[39m
#> [90m 7[39m 2012-01-01 [90m00:06:00[39m
#> [90m 8[39m 2012-01-01 [90m00:07:00[39m
#> [90m 9[39m 2012-01-01 [90m00:08:00[39m
#> [90m10[39m 2012-01-01 [90m00:09:00[39m
#> [90m# … with 86,390 more rows[39m

# Second series for 2 minutes
create_series('2011-01-01 12:10:00' ~ '2011-01-01 12:12:00', 's')

#> # A time tibble: 121 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2011-01-01 [90m12:10:00[39m
#> [90m 2[39m 2011-01-01 [90m12:10:01[39m
#> [90m 3[39m 2011-01-01 [90m12:10:02[39m
#> [90m 4[39m 2011-01-01 [90m12:10:03[39m
#> [90m 5[39m 2011-01-01 [90m12:10:04[39m
#> [90m 6[39m 2011-01-01 [90m12:10:05[39m
#> [90m 7[39m 2011-01-01 [90m12:10:06[39m
#> [90m 8[39m 2011-01-01 [90m12:10:07[39m
#> [90m 9[39m 2011-01-01 [90m12:10:08[39m
#> [90m10[39m 2011-01-01 [90m12:10:09[39m
#> [90m# … with 111 more rows[39m

# Date class
create_series(~'2013', 'day', class = "Date")

#> # A time tibble: 365 x 1
#> [90m# Index: date[39m
#>    date      
#>    [3m[90m<date>[39m[23m    
#> [90m 1[39m 2013-01-01
#> [90m 2[39m 2013-01-02
#> [90m 3[39m 2013-01-03
#> [90m 4[39m 2013-01-04
#> [90m 5[39m 2013-01-05
#> [90m 6[39m 2013-01-06
#> [90m 7[39m 2013-01-07
#> [90m 8[39m 2013-01-08
#> [90m 9[39m 2013-01-09
#> [90m10[39m 2013-01-10
#> [90m# … with 355 more rows[39m

# yearmon class
create_series(~'2013', 'month', class = "yearmon")

#> # A time tibble: 12 x 1
#> [90m# Index: date[39m
#>    date         
#>    [3m[90m<S3: yearmon>[39m[23m
#> [90m 1[39m Jan 2013     
#> [90m 2[39m Feb 2013     
#> [90m 3[39m Mar 2013     
#> [90m 4[39m Apr 2013     
#> [90m 5[39m May 2013     
#> [90m 6[39m Jun 2013     
#> [90m 7[39m Jul 2013     
#> [90m 8[39m Aug 2013     
#> [90m 9[39m Sep 2013     
#> [90m10[39m Oct 2013     
#> [90m11[39m Nov 2013     
#> [90m12[39m Dec 2013     

# hms class. time_formula specified as HH:MM:SS here
create_series('00:00:00' ~ '12:00:00', 'second' , class = "hms")

#> # A time tibble: 43,201 x 1
#> [90m# Index: date[39m
#>    date  
#>    [3m[90m<time>[39m[23m
#> [90m 1[39m 00[90m'[39m00[90m"[39m
#> [90m 2[39m 00[90m'[39m01[90m"[39m
#> [90m 3[39m 00[90m'[39m02[90m"[39m
#> [90m 4[39m 00[90m'[39m03[90m"[39m
#> [90m 5[39m 00[90m'[39m04[90m"[39m
#> [90m 6[39m 00[90m'[39m05[90m"[39m
#> [90m 7[39m 00[90m'[39m06[90m"[39m
#> [90m 8[39m 00[90m'[39m07[90m"[39m
#> [90m 9[39m 00[90m'[39m08[90m"[39m
#> [90m10[39m 00[90m'[39m09[90m"[39m
#> [90m# … with 43,191 more rows[39m

# Subsecond series
create_series('2013' ~ '2013-01-01 00:00:01', period = "10 millisec")

#> # A time tibble: 101 x 1
#> [90m# Index: date[39m
#>    date               
#>    [3m[90m<dttm>[39m[23m             
#> [90m 1[39m 2013-01-01 [90m00:00:00[39m
#> [90m 2[39m 2013-01-01 [90m00:00:00[39m
#> [90m 3[39m 2013-01-01 [90m00:00:00[39m
#> [90m 4[39m 2013-01-01 [90m00:00:00[39m
#> [90m 5[39m 2013-01-01 [90m00:00:00[39m
#> [90m 6[39m 2013-01-01 [90m00:00:00[39m
#> [90m 7[39m 2013-01-01 [90m00:00:00[39m
#> [90m 8[39m 2013-01-01 [90m00:00:00[39m
#> [90m 9[39m 2013-01-01 [90m00:00:00[39m
#> [90m10[39m 2013-01-01 [90m00:00:00[39m
#> [90m# … with 91 more rows[39m

milli <- create_series('2013' ~ '2013-01-01 00:00:01', period = ".1 sec")
# Check that 'milli' is correct by running:
# options("digits.secs" = 4)
# options("digits" = 18)
# milli$date
# as.numeric(milli$date)