A simple wrapper of lubridate::floor_date()

This is a thin wrapper around a lubridate::floor_date() that works for hms, yearmon, and yearqtr classes as well.

floor_index(x, unit = "seconds")

Arguments

x
a vector of date-time objects

unit
a character string specifying a time unit or a multiple of a unit to be rounded to. Valid base units are second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year. Arbitrary unique English abbreviations as in the period() constructor are allowed. Rounding to multiple of units (except weeks) is supported.

See also

Examples

data(FB)
dplyr::mutate(FB, date2 = floor_index(date, "year"))

#> [90m# A tibble: 1,008 x 9[39m
#>    symbol date        open  high   low close    volume adjusted date2     
#>    [3m[90m<chr>[39m[23m  [3m[90m<date>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<date>[39m[23m    
#> [90m 1[39m FB     2013-01-02  27.4  28.2  27.4  28    69[4m8[24m[4m4[24m[4m6[24m400     28   2013-01-01
#> [90m 2[39m FB     2013-01-03  27.9  28.5  27.6  27.8  63[4m1[24m[4m4[24m[4m0[24m600     27.8 2013-01-01
#> [90m 3[39m FB     2013-01-04  28.0  28.9  27.8  28.8  72[4m7[24m[4m1[24m[4m5[24m400     28.8 2013-01-01
#> [90m 4[39m FB     2013-01-07  28.7  29.8  28.6  29.4  83[4m7[24m[4m8[24m[4m1[24m800     29.4 2013-01-01
#> [90m 5[39m FB     2013-01-08  29.5  29.6  28.9  29.1  45[4m8[24m[4m7[24m[4m1[24m300     29.1 2013-01-01
#> [90m 6[39m FB     2013-01-09  29.7  30.6  29.5  30.6 104[4m7[24m[4m8[24m[4m7[24m700     30.6 2013-01-01
#> [90m 7[39m FB     2013-01-10  30.6  31.5  30.3  31.3  95[4m3[24m[4m1[24m[4m6[24m400     31.3 2013-01-01
#> [90m 8[39m FB     2013-01-11  31.3  32.0  31.1  31.7  89[4m5[24m[4m9[24m[4m8[24m000     31.7 2013-01-01
#> [90m 9[39m FB     2013-01-14  32.1  32.2  30.6  31.0  98[4m8[24m[4m9[24m[4m2[24m800     31.0 2013-01-01
#> [90m10[39m FB     2013-01-15  30.6  31.7  29.9  30.1 173[4m2[24m[4m4[24m[4m2[24m600     30.1 2013-01-01
#> [90m# … with 998 more rows[39m

time_test <- create_series('00:00:00'~'12:00:00',
                           '1 minute', class = "hms")

dplyr::mutate(time_test, date2 = floor_index(date, "hour"))

#> # A time tibble: 721 x 2
#> [90m# Index: date[39m
#>    date   date2 
#>    [3m[90m<time>[39m[23m [3m[90m<time>[39m[23m
#> [90m 1[39m 00[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 2[39m 01[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 3[39m 02[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 4[39m 03[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 5[39m 04[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 6[39m 05[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 7[39m 06[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 8[39m 07[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m 9[39m 08[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m10[39m 09[90m'[39m00[90m"[39m 00[90m'[39m00[90m"[39m
#> [90m# … with 711 more rows[39m