Create a rolling version of any function

rollify returns a rolling version of the input function, with a rolling window specified by the user.

rollify(.f, window = 1, unlist = TRUE, na_value = NULL)

Arguments

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get_attr() to extract named attributes. If a component is not present, the value of .default will be returned.

window
The window size to roll over

unlist
If the function returns a single value each time it is called, use unlist = TRUE. If the function returns more than one value, or a more complicated object (like a linear model), use unlist = FALSE to create a list-column of the rolling results.

na_value
A default value for the NA values at the beginning of the roll.

Details

The intented use of rollify is to turn a function into a rolling version of itself for use inside of a call to dplyr::mutate(), however it works equally as well when called from purrr::map().

Because of it's intended use with dplyr::mutate(), rollify always returns output with the same length of the input, aligned right, and filled with NA unless otherwise specified by na_value.

The form of the .f argument is the same as the form that can be passed to purrr::map(). Use .x or . to refer to the first object to roll over, and .y to refer to the second object if required. The examples explain this further.

If optional arguments to the function are required, specify them in the call to rollify, and not in the call to the rolling version of the function. See the examples for more details.

See also

purrr::safely, purrr::possibly

Examples

# Rolling mean --------------------------------------------------------------

data(FB)

# Turn the normal mean function into a rolling mean with a 5 row window
mean_roll_5 <- rollify(mean, window = 5)

dplyr::mutate(FB,
       normal_mean  = mean(adjusted),
       rolling_mean = mean_roll_5(adjusted))

#> # A tibble: 1,008 x 10
#>    symbol       date  open  high   low close    volume adjusted normal_mean
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>       <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00    77.51306
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77    77.51306
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76    77.51306
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42    77.51306
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06    77.51306
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59    77.51306
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30    77.51306
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72    77.51306
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95    77.51306
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10    77.51306
#> # ... with 998 more rows, and 1 more variables: rolling_mean <dbl>

# There's nothing stopping you from combining multiple rolling functions with
# different window sizes in the same mutate call
mean_roll_10 <- rollify(mean, window = 10)

dplyr::mutate(FB,
       rolling_mean_5  = mean_roll_5(adjusted),
       rolling_mean_10 = mean_roll_10(adjusted))

#> # A tibble: 1,008 x 10
#>    symbol       date  open  high   low close    volume adjusted rolling_mean_5
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>          <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00             NA
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77             NA
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76             NA
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42             NA
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06         28.602
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59         29.120
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30         29.826
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72         30.418
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95         30.724
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10         30.932
#> # ... with 998 more rows, and 1 more variables: rolling_mean_10 <dbl>

# Functions with multiple args and optional args ----------------------------

# With 2 args, use the purrr syntax of ~ and .x, .y
# Rolling correlation example
cor_roll <- rollify(~cor(.x, .y), window = 5)

dplyr::mutate(FB, running_cor = cor_roll(adjusted, open))

#> # A tibble: 1,008 x 9
#>    symbol       date  open  high   low close    volume adjusted running_cor
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>       <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00          NA
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77          NA
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76          NA
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42          NA
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06   0.7486236
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59   0.8051979
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30   0.8593871
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72   0.8838185
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95   0.6669151
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10   0.3786923
#> # ... with 998 more rows

# With >2 args, create an anonymous function with >2 args or use
# the purrr convention of ..1, ..2, ..3 to refer to the arguments
avg_of_avgs <- rollify(function(x, y, z) {
                         (mean(x) + mean(y) + mean(z)) / 3
                       },
                       window = 10)

# Or
avg_of_avgs <- rollify(~(mean(..1) + mean(..2) + mean(..3)) / 3,
                       window = 10)

dplyr::mutate(FB, avg_of_avgs = avg_of_avgs(open, high, low))

#> # A tibble: 1,008 x 9
#>    symbol       date  open  high   low close    volume adjusted avg_of_avgs
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>       <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400    28.00          NA
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77          NA
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76          NA
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42          NA
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06          NA
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59          NA
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30          NA
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72          NA
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95          NA
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10    29.68067
#> # ... with 998 more rows

# Optional arguments MUST be passed at the creation of the rolling function
# Only data arguments that are "rolled over" are allowed when calling the
# rolling version of the function
FB$adjusted[1] <- NA

roll_mean_na_rm <- rollify(~mean(.x, na.rm = TRUE), window = 5)

dplyr::mutate(FB, roll_mean = roll_mean_na_rm(adjusted))

#> # A tibble: 1,008 x 9
#>    symbol       date  open  high   low close    volume adjusted roll_mean
#>     <chr>     <date> <dbl> <dbl> <dbl> <dbl>     <dbl>    <dbl>     <dbl>
#>  1     FB 2013-01-02 27.44 28.18 27.42 28.00  69846400       NA        NA
#>  2     FB 2013-01-03 27.88 28.47 27.59 27.77  63140600    27.77        NA
#>  3     FB 2013-01-04 28.01 28.93 27.83 28.76  72715400    28.76        NA
#>  4     FB 2013-01-07 28.69 29.79 28.65 29.42  83781800    29.42        NA
#>  5     FB 2013-01-08 29.51 29.60 28.86 29.06  45871300    29.06   28.7525
#>  6     FB 2013-01-09 29.67 30.60 29.49 30.59 104787700    30.59   29.1200
#>  7     FB 2013-01-10 30.60 31.45 30.28 31.30  95316400    31.30   29.8260
#>  8     FB 2013-01-11 31.28 31.96 31.10 31.72  89598000    31.72   30.4180
#>  9     FB 2013-01-14 32.08 32.21 30.62 30.95  98892800    30.95   30.7240
#> 10     FB 2013-01-15 30.64 31.71 29.88 30.10 173242600    30.10   30.9320
#> # ... with 998 more rows

# Returning multiple values -------------------------------------------------

data(FB)

# If the function returns >1 value, set the `unlist = FALSE` argument
# Running 5 number summary
summary_roll <- rollify(summary, window = 5, unlist = FALSE)

FB_summarised <- dplyr::mutate(FB, summary_roll = summary_roll(adjusted))
FB_summarised$summary_roll[[5]]

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>   27.77   28.51   28.91   28.75   29.15   29.42       1 

# Rolling regressions -------------------------------------------------------

# Extending an example from R 4 Data Science on "Many Models".
# For each country in the gapminder data, calculate a linear regression
# every 5 periods of lifeExp ~ year
library(gapminder)

# Rolling regressions are easy to implement
lm_roll <- rollify(~lm(.x ~ .y), window = 5, unlist = FALSE)

gapminder %>%
  dplyr::group_by(country) %>%
  dplyr::mutate(rolling_lm = lm_roll(lifeExp, year))

#> # A tibble: 1,704 x 7
#> # Groups:   country [142]
#>        country continent  year lifeExp      pop gdpPercap rolling_lm
#>         <fctr>    <fctr> <int>   <dbl>    <int>     <dbl>     <list>
#>  1 Afghanistan      Asia  1952  28.801  8425333  779.4453  <lgl [1]>
#>  2 Afghanistan      Asia  1957  30.332  9240934  820.8530  <lgl [1]>
#>  3 Afghanistan      Asia  1962  31.997 10267083  853.1007  <lgl [1]>
#>  4 Afghanistan      Asia  1967  34.020 11537966  836.1971  <lgl [1]>
#>  5 Afghanistan      Asia  1972  36.088 13079460  739.9811   <S3: lm>
#>  6 Afghanistan      Asia  1977  38.438 14880372  786.1134   <S3: lm>
#>  7 Afghanistan      Asia  1982  39.854 12881816  978.0114   <S3: lm>
#>  8 Afghanistan      Asia  1987  40.822 13867957  852.3959   <S3: lm>
#>  9 Afghanistan      Asia  1992  41.674 16317921  649.3414   <S3: lm>
#> 10 Afghanistan      Asia  1997  41.763 22227415  635.3414   <S3: lm>
#> # ... with 1,694 more rows

# Rolling with groups -------------------------------------------------------

# One of the most powerful things about this is that it works with
# groups since `mutate` is being used
data(FANG)
FANG <- FANG %>%
  dplyr::group_by(symbol)

mean_roll_3 <- rollify(mean, window = 3)

FANG %>%
  dplyr::mutate(mean_roll = mean_roll_3(adjusted)) %>%
  dplyr::slice(1:5)

#> # A tibble: 20 x 9
#> # Groups:   symbol [4]
#>    symbol       date     open     high      low    close   volume  adjusted
#>     <chr>     <date>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>     <dbl>
#>  1   AMZN 2013-01-02 256.0800 258.1000 253.2600 257.3100  3271000 257.31000
#>  2   AMZN 2013-01-03 257.2700 260.8800 256.3700 258.4800  2750900 258.48001
#>  3   AMZN 2013-01-04 257.5800 259.8000 256.6500 259.1500  1874200 259.14999
#>  4   AMZN 2013-01-07 262.9700 269.7300 262.6700 268.4600  4910000 268.45999
#>  5   AMZN 2013-01-08 267.0700 268.9800 263.5700 266.3800  3010700 266.38000
#>  6     FB 2013-01-02  27.4400  28.1800  27.4200  28.0000 69846400  28.00000
#>  7     FB 2013-01-03  27.8800  28.4700  27.5900  27.7700 63140600  27.77000
#>  8     FB 2013-01-04  28.0100  28.9300  27.8300  28.7600 72715400  28.76000
#>  9     FB 2013-01-07  28.6900  29.7900  28.6500  29.4200 83781800  29.42000
#> 10     FB 2013-01-08  29.5100  29.6000  28.8600  29.0600 45871300  29.06000
#> 11   GOOG 2013-01-02 719.4212 727.0013 716.5512 723.2512  5101500 361.26435
#> 12   GOOG 2013-01-03 724.9313 731.9312 720.7212 723.6713  4653700 361.47415
#> 13   GOOG 2013-01-04 729.3412 741.4713 727.6812 737.9713  5547600 368.61701
#> 14   GOOG 2013-01-07 735.4513 739.3813 730.5812 734.7513  3323800 367.00863
#> 15   GOOG 2013-01-08 735.5413 736.3013 724.4312 733.3012  3364700 366.28433
#> 16   NFLX 2013-01-02  95.2100  95.8100  90.6900  92.0100 19431300  13.14429
#> 17   NFLX 2013-01-03  91.9700  97.9200  91.5300  96.5900 27912500  13.79857
#> 18   NFLX 2013-01-04  96.5400  97.7100  95.5400  95.9800 17761100  13.71143
#> 19   NFLX 2013-01-07  96.3900 101.7500  96.1200  99.2000 45550400  14.17143
#> 20   NFLX 2013-01-08 100.0100 100.9900  96.8000  97.1600 24714900  13.88000
#> # ... with 1 more variables: mean_roll <dbl>